### 2019-07-07 #1550 @whedon
[PRE REVIEW]: bbsAssistant: An R package for downloading and handling data and information from the North American Breeding Bird Survey.
**Submitting author:** @trashbirdecology (<a href="http://orcid.org/0000-0002-0896-5099">Jessica Burnett</a>)
**Repository:** <a href="https://github.com/TrashBirdEcology/bbsAssistant/" target ="_blank">github.com/trashbirdecology/bbsassistant</a>
**Version:** 0.0.1
