### 2019-07-07 #1550 @TrashBirdEcology
Oh no - so sorry about that!

https://github.com/TrashBirdEcology/bbsAssistant/

________________________________________________________________

Jessica L. Burnett, M.Sc.
Ph.D. Candidate
School of Natural Resources
University of Nebraska-Lincoln
+1.352.792.5425

GitHub <http://github.com/trashbirdecology>
CV/Resume
<https://github.com/TrashBirdEcology/cv/blob/master/burnett_extendedCV.pdf>
Nebraska Cooperative Fish & Wildlife Research Unit
<http://snr.unl.edu/necoopunit/>
ORC ID: 0000-0002-0896-5099 <https://orcid.org/0000-0002-0896-5099>
________________________________________________________________

<http://snr.unl.edu/necoopunit/>

On Sun, Jul 7, 2019 at 11:50 AM Lorena A. Barba <notifications@github.com>
wrote:

> 👋 @TrashBirdEcology <https://github.com/TrashBirdEcology> — the GitHub
> repo URL you provided with the submission is giving a 404. Can you post
> here the correct URL?
>
