### 2019-07-08 #1550 @kthyng
hi @TrashBirdEcology! I'll be your editor. I have a few comments and questions to start.

Please look through your paper and edit it so that your links look correct. I suspect you have a space between the [] and the () so they aren't rendering correctly.

Also please look through your readme — your "manual" link doesn't work, I think there is an image that isn't coming through ("coverage status") and I saw a typo on "speices".

Finally, have a look through our [list of reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit?usp=sharing) and select out a handful that you think could be good choices. I'll contact them to get two total. If you know other people who would be good reviewers (but aren't involved with the project), feel free to list them too. Thanks!
