### 2019-07-08 ea3d710
Fix typos
- Spacing b/w [] and ().
