+++ paper/paper.md
@@ -32,6 +32,6 @@ This package contains functions for downloading and munging data from the North 
 
 The BBS is an annual, roadside, volunteer-based survey of North American birds which began in 1966. It includes data from the continental United States, southern Canada, and more recently Alaska and northern Mexico.  More than 5,000 survey routes are tracked each June at the peak of the nesting season by experienced birders. Each route is aproximately 24.5 miles long, has 50 stops each located at 0.5 mile intervals. At each stop, the observer conducts a three-minute point count of all the birds heard or seen within 0.25 miles [@sauer1997north]. 
 
-The North American BBS [website] (https://www.pwrc.usgs.gov/bbs/index.cfm) produces an index of relative abundance and not a complete count of breeding bird populations. BBS data can be used to make continental- or regional-scale maps of relative abundance of breeding bird species; analyze population change on survey routes of some species over time to produce regional and continental estimates of trends; guild population trends and perform community analysis (e.g., neotropical migrant birds). For more information on how the BBS data can be used go to the BBS [website] (https://www.mbr-pwrc.usgs.gov/bbs/genintro.html).
+The North American BBS [website](https://www.pwrc.usgs.gov/bbs/index.cfm) produces an index of relative abundance and not a complete count of breeding bird populations. BBS data can be used to make continental- or regional-scale maps of relative abundance of breeding bird species; analyze population change on survey routes of some species over time to produce regional and continental estimates of trends; guild population trends and perform community analysis (e.g., neotropical migrant birds). For more information on how the BBS data can be used go to the BBS [website](https://www.mbr-pwrc.usgs.gov/bbs/genintro.html).
 
 # References
