+++ R/get_speciesList.R
@@ -1,6 +1,6 @@
 # Read in list of species names, from SpeciesList.txt, and then extract list of where the data is kept
 #' @title Download species names
-#' @description Read in list of species names, from SpeciesList.txt, and then extract list of where the data is kept
+#' @description Read in list of species names, from SpeciesList.txt, and then extract list of where the data is kept. This function was borrowed from the function [rBBS::getSpNames()](https://github.com/oharar/rBBS/blob/master/R/getspnames.r).
 #' @param Dir ftp URL for directory with data files
 #' @return A dataframe
 #' @export get_speciesList
