### 2019-07-08 aa41c2b
Update paper.md
Add acknowledgements. Fix links.
