+++ paper/paper.md
@@ -28,10 +28,11 @@ affiliations:
 
 # Summary
 
-This package contains functions for downloading and munging data from the North American Breeding Bird Survey (BBS) FTP server [@pardieck2018north; @sauer2017first]. Although the FTP server provides a user-friendly interface for retrieving data, this package consolidates the efforts of the data user by automating downloading and decompression of .zip data files, downloading route-level information, and saving them as .feather files for speedy import from disk. Notably, the package contains a web-scraping function for retrieving population trend estimates and data credibility of the results of the BBS hierarchical analyses [@sauer2017north].  Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
+This package contains functions for downloading and munging data from the North American Breeding Bird Survey (BBS) FTP server [@pardieck2018north; @sauer2017first]. Although the FTP server provides a public interface for retrieving data and analysis results, this package consolidates the efforts of the data user by automating downloading and decompression of .zip data files, downloading route-level information, and saving them as .feather files for speedy import from disk. The data subsetting features of this package also allow the user to readily import and save to file only the data necessary for her purposes. Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
 
-The BBS is an annual, roadside, volunteer-based survey of North American birds which began in 1966. It includes data from the continental United States, southern Canada, and more recently Alaska and northern Mexico.  More than 5,000 survey routes are tracked each June at the peak of the nesting season by experienced birders. Each route is aproximately 24.5 miles long, has 50 stops each located at 0.5 mile intervals. At each stop, the observer conducts a three-minute point count of all the birds heard or seen within 0.25 miles [@sauer1997north]. 
+The BBS team uses hierarhical modelling techniques to generate population trend estimates [@sauer2017north] at various spatial scales [see the BBS results webpage](https://www.mbr-pwrc.usgs.gov/). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package contains two functions for retrieving the population trend estimates produced by @sauer2017north and the associated data credibility scores: a web-scraping function for obtaining current region and/or species-specific population trend estimates and data credibility scores via a supplied url, [`get_credibility_trends()`](https://github.com/TrashBirdEcology/bbsAssistant/blob/master/R/get_credibility_trends.R); and a function for the current and archived population trends estimates for *all* species and regions, [`get_analysis_results()`](https://github.com/TrashBirdEcology/bbsAssistant/blob/master/R/get_analysis_results.R). 
 
-The North American BBS [website](https://www.pwrc.usgs.gov/bbs/index.cfm) produces an index of relative abundance and not a complete count of breeding bird populations. BBS data can be used to make continental- or regional-scale maps of relative abundance of breeding bird species; analyze population change on survey routes of some species over time to produce regional and continental estimates of trends; guild population trends and perform community analysis (e.g., neotropical migrant birds). For more information on how the BBS data can be used go to the BBS [website](https://www.mbr-pwrc.usgs.gov/bbs/genintro.html).
+# Acknowledgements
+We thank the volunteer citizen scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. Some functions in this package were adapted from the [rBBS](github.com/oharar/rbbs) package and are mentioned in function source code as appicable.
 
 # References
