### 2019-07-08 78cf4ac
Clean up readme.
- Add glimpse() to some echoed output
- Remove coverage status img link
-
