+++ README.Rmd
@@ -10,7 +10,7 @@ bibliography: paper/paper.bib
 
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![Travis build status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)
-[![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master)
+<!-- [![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
 
 
 # Function Manual
@@ -19,7 +19,7 @@ For function descriptions and examples please see the [manual](/man/bbsAssistant
 # Contributions
 If you would like to contribute, please submit a pr or email  (jessicaleighburnett at gmail). I am especially interested in having another set of eyes and hands to transfer and clean up the functions listed in [Issue 1](https://github.com/TrashBirdEcology/bbsAssistant/issues/1).
 
-# Brief Overview of bbsAssistant Features
+# Overview of bbsAssistant Features
 
 ```{r  setup, include=FALSE, echo=FALSE, warning=FALSE, message=FALSE}
 knitr::opts_chunk$set(echo=TRUE, warning = FALSE, message=FALSE)
@@ -72,7 +72,7 @@ Next, we can download the BBS route-level geographic information and metadata, a
 ```{r, echo=TRUE}
 routes <- get_routeInfo() # retrieve route-level data
 flBBS <- left_join(flBBS, routes) # merge route-level data to bird count data
-glimpse(flBBS)
+glimpse(flBBS %>% dplyr::select(aou, year, route, statenum, countrynum, stoptotal, latitude, longitude))
 ```
 
 If we wish to save these data to file, we can do so by saving as **.feather*s, a compressed file formatted for use in R.
@@ -87,19 +87,13 @@ export_bbsFeathers(dataIn = flBBS,
 
 If the BBS data was downloaded previously and saved as .feather, we can import it using `import_bbsFeathers`. The code below is particularly useful if you are importing multiple files (e.g., multiple states)
 ```{r import_bbsFeathers, echo=TRUE}
-feathers <- NULL
+
 (featherNames <- list.files(bbsDir, pattern = ".feather"))
 featherNames <- str_c("/", featherNames) #add separator
-for (i in 1:length(featherNames)) {
-  feather <- NULL
-  feather <- import_bbsFeathers(newDir  = bbsDir,
-                              filename = featherNames[i])
-  feathers <- rbind(feathers, feather)
-  rm(feather)
-}
 
-# you will notice the imported data slightly differs from the original bbs data in that # cols is far fewer
-ncol(feathers); ncol(flBBS)
+feather <- import_bbsFeathers(newDir  = bbsDir,
+                              filename = featherNames)
+glimpse(feather) # Notice that the data imported from disk (feathers) differs from the original BBS data in that the # of columns is fewer (9 and 12 columns, respectively).
 ```
 
 ### Option for downloading ALL BBS data
@@ -119,17 +113,17 @@ for(i in 1:length(regionFileName)){
             # arguments for get_bbsData()
             file = regionFileName[i],
             dir =  "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/States/",
-            year = NULL,
-            aou = NULL,
-            countrynum = NULL,
-            states = NULL,
[…]
