### 2019-07-08 #1550 @TrashBirdEcology
@kthyng Thank you for this quick feedback.

I have addressed the issues you mentioned in man and readme.

I recommend @ethanwhite and @jhollist as potential reviewers.
Please note that I have a mild conflict of interest with potential reviewer @ha0ye

Thanks!
