### 2019-07-08 c02ebfa
Update function documentation + manual
- Spelling in readme + fxns
- Update link to images in readme
- Re-build manual
- Fix .bib entry typo
