+++ R/get_analysis_results.R
@@ -1,6 +1,6 @@
 #' @title Get annual abundance indices from BBS analyses. 
-#' @description Downloads all species-regions annual abundances indices from various .csv files on the BBS analysis results page. See <https://www.mbr-pwrc.usgs.gov/bbs/BBSAnalysis_Results_Metadata_1966_2017_v3.xml> for metadata for analysis results and trend estiamtes. 
-#' @param analysis Default = "trend.ests". Must be one of :c("annual.inds.2016", "trend.ests","core.trend.revised.2017", "core.2017","expanded.2017","core.twedt.revised.2017")
+#' @description Downloads all species-regions annual abundances indices from various .csv files on the BBS analysis results page. See [results metadata](https://www.mbr-pwrc.usgs.gov/bbs/BBSAnalysis_Results_Metadata_1966_2017_v3.xml) for analysis results and trend estimates. 
+#' @param analysis Default = "trend.ests". Must be one of c("annual.inds.2016", "trend.ests","core.trend.revised.2017", "core.2017","expanded.2017","core.twedt.revised.2017")
 #' @param url Web link to the .csv file for annual abundance indices. Default = NULL.
 #' @export get_analysis_results
 
