+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{sauer2017north,
   title={The North American breeding bird survey, results and analysis 1966-2015},
-  author={Sauer, John and Niven, Daniel and Hines, James and Ziolkowski Jr, David and Pardieck, Keith L and Fallon, JE and Link, William},
+  author={Sauer, JR and Niven, D and Hines, J and Ziolkowski Jr, David and Pardieck, KL and Fallon, JE and Link, William},
   year={2017}
 }
 
@@ -26,7 +26,7 @@
 
 @article{sauer1997north,
   title={The North American breeding bird survey results and analysis, version 96.4},
-  author={Sauer, Jo R and Hines, JE and Gough, G and Thomas, I and Peterjohn, BG},
+  author={Sauer, JR and Hines, JE and Gough, G and Thomas, I and Peterjohn, BG},
   journal={Patuxent Wildlife Research Center, Laurel, Maryland},
   year={1997}
 }
