### 2019-07-08 #1550 @kthyng
@TrashBirdEcology Great!

I see that @jhollist has recently done a review for us, so I will try to find someone else.

@ethanwhite and @r-barnes are you interested and available to do a review on this JOSS submission? We ask for reviews within about 3 weeks, and you can get more information [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
