### 2019-07-09 5e812ed
Update export_bbsFeathers.R
Add logical for newDir ending in "/". Append / if last chr isnt == "/"
