### 2019-07-09 5fe7b58
Update license, descrip whitespace
