### 2019-07-09 dcc58ca
Update manual + add logo draft (needs higher DPI)
