+++ README.Rmd
@@ -11,10 +11,11 @@ bibliography: paper/paper.bib
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![Travis build status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)
 <!-- [![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
+ # bbsAssistant <img src="man/figures/logo.png" align="right" height=140/>
 
 
 # Function Manual
-For function descriptions and examples please see the [manual](/man/bbsAssistant_0.0.0.9000.pdf). 
+For function descriptions and examples please see the [manual](https://github.com/trashbirdecology/bbsassistant/bbsAssistant_0.0.1.pdf). 
 
 # Contributions
 If you would like to contribute, please submit a pr or email  (jessicaleighburnett at gmail). I am especially interested in having another set of eyes and hands to transfer and clean up the functions listed in [Issue 1](https://github.com/TrashBirdEcology/bbsAssistant/issues/1).
@@ -65,6 +66,7 @@ regionFileName <- regions$zipFileName %>% na.omit()
 
 Once we have one or more region filenames, we can use function `get_bbsData` to download the .zip file to a temporary folder (unless otherwise specified), and *import* the temp file to R object. The R object, flBBS, contains the raw BBS count data.
 ```{r, echo=TRUE}
+# check to see whether the data we need is already on file. if not, download it.
 flBBS <- get_bbsData(file=regionFileName.use)
 ```
 
@@ -87,7 +89,6 @@ export_bbsFeathers(dataIn = flBBS,
 
 If the BBS data was downloaded previously and saved as .feather, we can import it using `import_bbsFeathers`. The code below is particularly useful if you are importing multiple files (e.g., multiple states)
 ```{r import_bbsFeathers, echo=TRUE}
-
 (featherNames <- list.files(bbsDir, pattern = ".feather"))
 featherNames <- str_c("/", featherNames) #add separator
 
