### 2019-07-09 b64bc89
Update logo
