+++ README.md
@@ -106,14 +106,14 @@ glimpse(flBBS %>% dplyr::select(aou, year, route, statenum, countrynum, stoptota
 
     ## Observations: 139,735
     ## Variables: 8
-    ## $ aou        <int> 1840, 2000, 2010, 2890, 3100, 3131, 3160, 3390, 343...
-    ## $ year       <int> 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967, 196...
-    ## $ route      <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
-    ## $ statenum   <int> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25,...
-    ## $ countrynum <int> 840, 840, 840, 840, 840, 840, 840, 840, 840, 840, 8...
-    ## $ stoptotal  <dbl> 3, 1, 1, 37, 1, 1, 7, 1, 1, 1, 1, 1, 2, 11, 5, 4, 2...
-    ## $ latitude   <dbl> 30.92918, 30.92918, 30.92918, 30.92918, 30.92918, 3...
-    ## $ longitude  <dbl> -87.40794, -87.40794, -87.40794, -87.40794, -87.407...
+    ## $ aou        <int> 1840, 2000, 2010, 2890, 3100, 3131, 3160, 3390, 3430,…
+    ## $ year       <int> 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967,…
+    ## $ route      <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
+    ## $ statenum   <int> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 2…
+    ## $ countrynum <int> 840, 840, 840, 840, 840, 840, 840, 840, 840, 840, 840…
+    ## $ stoptotal  <dbl> 3, 1, 1, 37, 1, 1, 7, 1, 1, 1, 1, 1, 2, 11, 5, 4, 2, …
+    ## $ latitude   <dbl> 30.92918, 30.92918, 30.92918, 30.92918, 30.92918, 30.…
+    ## $ longitude  <dbl> -87.40794, -87.40794, -87.40794, -87.40794, -87.40794…
 
 If we wish to save these data to file, we can do so by saving as
 \*\*.feather\*s, a compressed file formatted for use in R.
@@ -147,15 +147,15 @@ glimpse(feather) # Notice that the data imported from disk (feathers) differs fr
 
     ## Observations: 139,735
     ## Variables: 9
-    ## $ year       <int> 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967, 196...
-    ## $ countrynum <int> 840, 840, 840, 840, 840, 840, 840, 840, 840, 840, 8...
-    ## $ statenum   <int> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25,...
-    ## $ route      <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
-    ## $ bcr        <int> 27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27,...
-    ## $ latitude   <dbl> 30.92918, 30.92918, 30.92918, 30.92918, 30.92918, 3...
-    ## $ longitude  <dbl> -87.40794, -87.40794, -87.40794, -87.40794, -87.407...
[…]
