### 2019-07-09 42f961a
Add original png logo and update README with resized.
logo_draft = original
logo  = usethis::use_logo() auto resize
