### 2019-07-11 da443c7
Update README + silence .tex and .log rendered
