+++ README.md
@@ -1,8 +1,5 @@
 <!-- PLEASE DO NOT EDIT README.md BY HAND. Please render README.rmd! -->
-
-[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis
-build
-status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>
+<!-- [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis build status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/> -->
 <!-- [![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
 
 Function Manual
@@ -106,14 +103,14 @@ glimpse(flBBS %>% dplyr::select(aou, year, route, statenum, countrynum, stoptota
 
     ## Observations: 139,735
     ## Variables: 8
-    ## $ aou        <int> 1840, 2000, 2010, 2890, 3100, 3131, 3160, 3390, 3430,…
-    ## $ year       <int> 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967,…
-    ## $ route      <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
-    ## $ statenum   <int> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 2…
-    ## $ countrynum <int> 840, 840, 840, 840, 840, 840, 840, 840, 840, 840, 840…
-    ## $ stoptotal  <dbl> 3, 1, 1, 37, 1, 1, 7, 1, 1, 1, 1, 1, 2, 11, 5, 4, 2, …
-    ## $ latitude   <dbl> 30.92918, 30.92918, 30.92918, 30.92918, 30.92918, 30.…
-    ## $ longitude  <dbl> -87.40794, -87.40794, -87.40794, -87.40794, -87.40794…
+    ## $ aou        <int> 1840, 2000, 2010, 2890, 3100, 3131, 3160, 3390, 343...
+    ## $ year       <int> 1967, 1967, 1967, 1967, 1967, 1967, 1967, 1967, 196...
+    ## $ route      <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
+    ## $ statenum   <int> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25,...
+    ## $ countrynum <int> 840, 840, 840, 840, 840, 840, 840, 840, 840, 840, 8...
+    ## $ stoptotal  <dbl> 3, 1, 1, 37, 1, 1, 7, 1, 1, 1, 1, 1, 2, 11, 5, 4, 2...
+    ## $ latitude   <dbl> 30.92918, 30.92918, 30.92918, 30.92918, 30.92918, 3...
+    ## $ longitude  <dbl> -87.40794, -87.40794, -87.40794, -87.40794, -87.407...
 
 If we wish to save these data to file, we can do so by saving as
[…]
