### 2019-07-16 0c37859
Update readme contributions section
Minor update
