### 2019-07-16 5ea153f
Silence cov badge .md
