### 2019-07-16 d1c9a6d
Fix tests + update travis YAML
add travis to ignore

Update all test files to generate objects outside the expectation calls
