+++ tests/testthat.R
@@ -6,3 +6,7 @@ library(gdata)
 library(xml2)
 
 test_check("bbsAssistant")
+
+
+# Run tests
+# testthat::test_package("bbsAssistant")
