+++ tests/testthat/test-get_regions.R
@@ -1,5 +1,9 @@
 context("get_regions")
 
+library(stringr)
+
+RegionCodes <- get_regions()
+
 test_that("Region data contains US and CA countries & the countryNum column name is assigned correctly.", {
     
     if(!exists('RegionCodes')) RegionCodes <- get_regions()
