+++ R/get_regions.R
@@ -53,23 +53,23 @@ get_regions <-
             fileEncoding = 'Latin1',
             ## WOULD LIKE A BETTER SOLUTION THAN SKIPPING X LINES
             skip = 168
-        ) %>% as_tibble()
+        ) %>% dplyr::as_tibble()
         
         
         ## WOULD LIKE A BETTER SOLUTION THAN SKIPPING X LINES
         readme.all = readme.all[1:62,]
         
         zipf = readme.all %>%
-            mutate(V1 =
+            dplyr::mutate(V1 =
                        stringr::str_replace(V1, "Comma Delimited", "")) %>%
-            mutate(zipFileName =
+            dplyr::mutate(zipFileName =
                        stringr::word(V1, 1)) %>%
-            mutate(stateName = str_squish(toupper(stringr::word(V1, 2:3, -1)))) %>%
+            dplyr::mutate(stateName = stringr::str_squish(toupper(stringr::word(V1, 2:3, -1)))) %>%
             dplyr::select(zipFileName, stateName)
         
         RegionCodes <-
-            right_join(RegionCodes %>%  filter(countryNum != 484) %>%
-                           mutate(stateName = toupper(stateName)),
+            dplyr::right_join(RegionCodes %>%  dplyr::filter(countryNum != 484) %>%
+                                  dplyr::mutate(stateName = toupper(stateName)),
                        zipf)
         
         ## Returns a data frame with country code (#), state number (#), state/region name (alpha), and filename
