+++ tests/testthat/test-subsetSpeciesList.R
@@ -1,14 +1,18 @@
 context("test-subsetSpeciesList")
 
-test_that("Ensure species list subsetting works (by order + fam)",{
-    
-    spp <- get_speciesList()
+spp <- get_speciesList()
+t <- subset_speciesList(spp, fam.ind="Parulidae")
+t2 <- subset_speciesList(spp, order.ind="Passeriformes")
 
-    t <- subset_speciesList(spp, fam.ind="Parulidae")
+test_that("Ensure species list subsetting works (by family)",{
     
     expect_false("Parulidae" %in% t$family )
-  
-    t <- subset_speciesList(spp, order.ind="Passeriformes")
-    expect_false("Passeriformes" %in% t$order )
-    
 } )   
+    
+
+
+test_that("Ensure species list subsetting works (by order )",{
+    
+    expect_false("Passeriformes" %in% t2$order )
+    
+} ) 
