+++ .travis.yml
@@ -1,1 +1,8 @@
+# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+
 language: r
+sudo: false
+cache: packages
+
+after_success:
+    - Rscript -e 'covr::codecov()'
