### 2019-07-16 5c6f799
Update .travis.yml
Ignore vignettes, add codecov, specify email notifications to JLB
