+++ .travis.yml
@@ -3,6 +3,32 @@
 language: r
 sudo: false
 cache: packages
+warnings_are_errors: false
 
+# environment variables set for all builds
+env:
+  global:
+    - BIOC_USE_DEVEL="FALSE"  ## Use the current release version
+    - R_BUILD_ARGS="--no-build-vignettes --no-manual"
+    - R_CHECK_ARGS="--no-build-vignettes --no-manual --timings"  ## do not build vignettes or manual
+    - _R_CHECK_TIMINGS_="0"  ## get the timing information for the examples for all of your functions
+
+r:
+ - release
+
+# do not build vignettes...takes too long and times out on travis
+r_build_args: --no-build-vignettes --no-manual
+r_check_args: --no-build-vignettes --no-manual --timings
+
+# for codecov
+r_packages:
+  - covr
+
+# only report coverage for the release version
 after_success:
-    - Rscript -e 'covr::codecov()'
+  - test $TRAVIS_R_VERSION_STRING = 'release' && Rscript -e 'covr::codecov()'
+
+notifications:
+  email:
+    on_success: jessicaleighburnett@gmail.com
+    on_failure: jessicaleighburnett@gmail.com
