### 2019-07-16 92ae2e0
Revert "Update .travis.yml"
This reverts commit 5c6f799068fbb7533a718dde02841a696ce55516.
