### 2019-07-16 bc3e552
Update manual + travis
Major updates to travis.yml
