### 2019-07-16 de3e342
Silence necessary lines
