### 2019-07-16 5387145
Update depends
Specify package in every dependent function (ie. add pkg::function())
Update documentation

Hopefully editing these imports was what was preventing travis build..

Update manual and doucmentaiton
