+++ R/get_credibility_trends.R
@@ -4,61 +4,83 @@
 #' @export get_credibility_trends
 #' @importFrom magrittr %>%
 
-get_credibility_trends <- function(url = 'https://www.mbr-pwrc.usgs.gov/cgi-bin/atlasa15.pl?FLA&2&15&csrfmiddlewaretoken=3YKakk7LxT2ki6NSpl4mstudYCqdW02C'){
-  
-#Reading the HTML code from the specified website (example used == Kansas)
-webpage <- read_html(url)
-
-# Get colored dot credibility ratings
-# grab images from xml nodes
-    ## the credibility ratings (red, yellow, blue) are provided only as "images". 
-    ## we need to capture the image names to associate a credibility rating to the species-region combination
-dotList <- webpage %>% 
- html_nodes("img") %>% 
-html_attrs() %>%
-  as.list()
-
-# this is ugly, but works. Drop heading and summary images, make dataframe with indexes
-
-dotKey = data.frame(credibilityColor = c("Red", "Yellow", "Blue"),
-                    credibilityClass = c("important_deficiency", "deficiency", "no_deficiency"),
-                    credibilityNumber = c("2", "1", "0"))
-
-dotDF =    data.frame(credibilityNumber = unlist(dotList[4:length(dotList)])) %>%
-           mutate(credibilityNumber = substr(credibilityNumber, 6, 6)) %>%
-           left_join(dotKey, by = "credibilityNumber")
-    
-# Scrape trend data
-spp_credibility_trends <- webpage %>%
-  html_nodes("#maincontent") %>%
-  html_text() %>%
-  strsplit(split = "\n") %>%
-  unlist() %>%
-  as.list()
-
-# clean up trend data
-
-spp_credibility_trends[5:(length(spp_credibility_trends) - 1)] %>%
-unlist() %>% 
-  as_tibble() %>%
-  #mutate(value = gsub(" ", ".", value)) %>%
-  extract(value, c("Species", "N", "Trend_1966_2015",
-                   "CI_95_1966_2015", "Trend_2005_2015",
-                   "CI_95_2005_2015", "RA"), "(.............................)(....)(.........)(....................)(.........)(....................)(.........)",
-                convert=TRUE) %>%
-  mutate(CI_95_1966_2015 = gsub("\\(", "", CI_95_1966_2015),
-         CI_95_1966_2015 = gsub("\\)", "", CI_95_1966_2015),
-         CI_95_2005_2015 = gsub("\\(", "", CI_95_2005_2015),
-         CI_95_2005_2015 = gsub("\\)", "", CI_95_2005_2015),
-         Species = gsub("\\(all forms\\)", "", Species),
-         Species = gsub("\\(all f\\)", "", Species),
-         Species = gsub("\\(all fo\\)", "", Species),
-         Species = gsub("\\(all\\)", "", Species)) %>%
-  separate(CI_95_1966_2015, into = c("CI_2.5_1966_2015", "CI_97.5_1966_2015"), sep = ",") %>%
-  separate(CI_95_2005_2015, into = c("CI_2.5_2005_2015", "CI_97.5_2005_2015"), sep = ",") %>%
-  mutate(Species = gsub("\\(*","",Species)) %>%
-  cbind(dotDF) %>% trim() -> spp_credibility_trends
-
-  
-return(spp_credibility_trends)
+get_credibility_trends <-
+    function(url = 'https://www.mbr-pwrc.usgs.gov/cgi-bin/atlasa15.pl?FLA&2&15&csrfmiddlewaretoken=3YKakk7LxT2ki6NSpl4mstudYCqdW02C') {
+        #Reading the HTML code from the specified website (example used == Kansas)
+        webpage <- xml2::read_html(url)
+        
+        # Get colored dot credibility ratings
[…]
