+++ tests/testthat/test-get_regions.R
@@ -1,6 +1,5 @@
 context("get_regions")
 
-library(stringr)
 
 RegionCodes <- get_regions()
 
