+++ R/subset_speciesList.R
@@ -17,17 +17,17 @@ subset_speciesList <- function(myData,
     ## Will update to remove by families and orders for the remove.fowl etc...
     # if("remove.fowl" %in% subset.by){
     #     myData <- myData %>%
-    #         filter(!aou %in% c(01290:01780))
+    #         dplyr::filter(!aou %in% c(01290:01780))
     # }
     #
     # if( "remove.shorebirds" %in% subset.by){
     #     myData <- myData %>%
-    #         filter(!aou %in% c(1290:1780))
+    #         dplyr::filter(!aou %in% c(1290:1780))
     # }
     #
     # if("remove.shoreWaderFowl" %in% subset.by){
     #     myData <- myData %>%
-    #         filter(!aou %in% c(00000:02880))
+    #         dplyr::filter(!aou %in% c(00000:02880))
     # }
     
     
@@ -35,20 +35,20 @@ subset_speciesList <- function(myData,
     ## Subset by taxonomic families
     if (!is.null(fam.ind)) {
         myData <- myData %>%
-            filter(!family %in% fam.ind)
+            dplyr::filter(!family %in% fam.ind)
     }
     
     
     ## Subset by taxonomic order
     if (!is.null(order.ind)) {
         myData <- myData %>%
-            filter(!order %in% order.ind)
+            dplyr::filter(!order %in% order.ind)
     }
     
     ## Subset by AOU specific numbers.
     if (!is.null(aou.ind)) {
         myData <- myData %>%
-            filter(!aou %in% aou.ind)
+            dplyr::filter(!aou %in% aou.ind)
     }
     
     return(myData)
