### 2019-07-16 eef9de8
Issue with get_credibility_trends
The function "get_credibility_trends" contains the function 'extract'. Need to determine what the source package is, beucae magrittr::extract() does not have argument "convert (logical)".  Please address this @lsw5077
