+++ DESCRIPTION
@@ -22,6 +22,7 @@ Imports:
     xml2, 
     feather,
     here,
+    glue, 
     magrittr
 Suggests:
     testthat (>= 2.1.0),
