### 2019-07-16 1c7f714
Update tests.
@lsw5077  when you identify/fix the issue with get_credibility trends, will you please unsilence the tests in tests/testthat/test-get_credibility-and-trends.R ?
