+++ tests/testthat.R
@@ -1,12 +1,8 @@
-library(testthat)
-library(bbsAssistant)
-library(tidyverse)
-library(rvest)
-library(gdata)
-library(xml2)
-
-test_check("bbsAssistant")
+library("testthat")
+library("bbsAssistant")
 
 
 # Run tests
-# testthat::test_package("bbsAssistant")
+testthat::test_package("bbsAssistant")
+
+
