### 2019-07-16 c3e6aeb
Update .travis.yml
Update code cov integration
