+++ .travis.yml
@@ -6,8 +6,7 @@ cache: packages
 warnings_are_errors: false
 
 after_success:
-    - Rscript -e 'covr::codecov()'
-
+    - Rscript -e 'library(covr); codecov()'
 email:
     recipients:
       - jessicaleighburnett@gmail.com
