### 2019-07-16 ede7ae7
Update testthat.R
