### 2019-07-16 91fc640
Remove library loads from readme
library(bbsAssistant) only
