+++ .travis.yml
@@ -28,6 +28,7 @@ r_check_args: --no-build-vignettes --no-manual --timings
 # for codecov
 r_packages:
   - covr
+  - testthat
 
 after_success:
   - Rscript -e 'library(covr); codecov()'
