### 2019-07-16 f82f84c
README minor suggestions.
