### 2019-07-16 607d40e
comment out install_github("trashbirdecology/bbsAssistant")
