### 2019-07-16 c6815d2
Update testthat.R
