+++ tests/testthat.R
@@ -1,9 +1,4 @@
 library("testthat")
 library("bbsAssistant")
 
-
-# Run tests
-testthat::test_dir(here::here('tests/testthat/'))
-
-
-
+test_check("bbsAssistant")
