### 2019-07-16 80b15d2
Input multiple regions with regionFileName example. This example wasn't in the original README.
