+++ NAMESPACE
@@ -12,3 +12,7 @@ export(import_bbsData)
 export(import_bbsFeathers)
 export(subset_speciesList)
 importFrom(magrittr,"%>%")
+importFrom(stats,family)
+importFrom(utils,download.file)
+importFrom(utils,read.csv)
+importFrom(utils,read.table)
