### 2019-07-16 e03d21f
Fix get_credibility_trends
It's not a nice fix, but it works for me....Forcing character classes to numeric is ugly....

Update documentation
