### 2019-07-16 2506c01
Update .travis.yml
Trying to debug travis... remove sudo = false
