+++ .travis.yml
@@ -1,7 +1,7 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: r
-sudo: false
+#sudo: false
 cache: packages
 warnings_are_errors: false
 
