+++ tests/testthat/test-get_regions.R
@@ -1,6 +1,5 @@
 context("get_regions")
 
-
 RegionCodes <- get_regions()
 
 test_that("Region data contains US and CA countries & the countryNum column name is assigned correctly.", {
