### 2019-07-17 ac82aad
Skip tests on travis.
All tests require local file writing and an internet connection. I don't thikn this is possible (or is difficult) on CI
