+++ tests/testthat/test-get_regions.R
@@ -1,23 +1,27 @@
 context("get_regions")
 
-RegionCodes <- get_regions()
 
-test_that("Region data contains US and CA countries & the countryNum column name is assigned correctly.", {
+test_that(
+  "Region data contains US and CA countries & the countryNum column name is assigned correctly.",
+  {
+    skip_on_travis()
+    RegionCodes <- get_regions()
+    if (!exists('RegionCodes'))
+      RegionCodes <- get_regions()
     
-    if(!exists('RegionCodes')) RegionCodes <- get_regions()
-  
-    expect_true(
-      124 %in% unique(RegionCodes$countryNum) &
-                   840 %in% unique(RegionCodes$countryNum)
-                )     
-    })
+    expect_true(124 %in% unique(RegionCodes$countryNum) &
+                  840 %in% unique(RegionCodes$countryNum))
+  }
+)
 
-test_that("Region data contains US and CA countries & the stateName column name is assigned correctly.", {
+test_that(
+  "Region data contains US and CA countries & the stateName column name is assigned correctly.",
+  {
+    skip_on_travis()
     
-    if(!exists('RegionCodes')) RegionCodes <- get_regions()
+    if (!exists('RegionCodes'))
+      RegionCodes <- get_regions()
     
-    expect_true(
-        "FLORIDA" %in% unique(RegionCodes$stateName) 
-    )     
-})
-
+    expect_true("FLORIDA" %in% unique(RegionCodes$stateName))
+  }
+)
