+++ tests/testthat/test-get_speciesList.R
@@ -1,14 +1,22 @@
 context("test-get_speciesList")
 
-spp <- get_speciesList()
 
-test_that("Ensure import of specieslist.txt is correct by checking col names", {
-
-    expect_true(all(
-        sort(names(spp))==c(
-        "aou","commonName","family",
-        "frenchCommonName", "genus","order",  
-        "scientificName","seq","species"))
-        )
-    
-})
+test_that("Ensure import of specieslist.txt is correct by checking col names",
+          {
+              skip_on_travis()
+              spp <- get_speciesList()
+              expect_true(all(
+                  sort(names(spp)) == c(
+                      "aou",
+                      "commonName",
+                      "family",
+                      "frenchCommonName",
+                      "genus",
+                      "order",
+                      "scientificName",
+                      "seq",
+                      "species"
+                  )
+              ))
+              
+          })
