+++ tests/testthat/test-subsetSpeciesList.R
@@ -1,16 +1,15 @@
 context("test-subsetSpeciesList")
 
+
+test_that("Ensure species list subsetting works (by family, then order)", {
+skip_on_travis()
 spp <- get_speciesList()
 t <- subset_speciesList(spp, fam.ind = "Parulidae")
-t2 <- subset_speciesList(spp, order.ind = "Passeriformes")
-
-test_that("Ensure species list subsetting works (by family)", {
     expect_false("Parulidae" %in% t$family)
-})
-
-
-
-test_that("Ensure species list subsetting works (by order )", {
+t2 <- subset_speciesList(spp, order.ind = "Passeriformes")
     expect_false("Passeriformes" %in% t2$order)
-    
+
+
 })
+
+
