### 2019-07-17 3c32c59
Update tests and licenses
