### 2019-07-17 dcc18d4
Updte documentation + license
