### 2019-07-17 3a07c92
Add vignettes, pkgdown website
