### 2019-07-17 e372df8
Update readme, move images ot man, add docs/, update readme
Also moved the files in /images/ to man/figures/...

Add .github documents, moved all non-traditional first-levels to that folder (e.g. contributing.md, code of conduct)

Update pkgdown website
