+++ README.md
@@ -1,44 +1,44 @@
-<!-- PLEASE DO NOT EDIT README.md BY HAND. Please render README.rmd! -->
+**bbsAssistant**: An R package for downloading and handling data and
+information from the North American Breeding Bird Survey.
+================
+2019-07-17
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
 <!-- Silence lines 12 - 13 when rendering .pdf.  -->
 
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis
 build
 status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>
-<!--[![Coverage
-status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
+![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)
+<!-- [![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
 
-Function Manual
-===============
+# Function Descriptions and Vignettes
 
-For function descriptions and examples please see the
-[manual](https://github.com/trashbirdecology/bbsassistant/bbsAssistant_0.0.1.pdf).
+For function descriptions please build the manual
+(`devtools::build_manual("bbsAssistant)`) and for an example build the
+vignette (`usethis::build_vignettes()`; or run
+`/vignettes/vignettes.Rmd`).
 
-Contributions
-=============
+# Contributions
 
-If you would like to contribute, please submit a pr, create an issue, or
-email us (jessicaleighburnett at gmail). Contributors must adhere to the
-[Code of Conduct](CODE_OF_CONDUCT.MD).
+To make a contribution visit the [contributions
+instructions](.github/CONTRIBUTIONS.MD). Contributors must adhere to the
+[Code of Conduct](.github/CODE_OF_CONDUCT.MD).
 
-Overview of bbsAssistant Features
-=================================
+# Brief Overview of bbsAssistant Functionality
 
-Installing package and loading dependencies
--------------------------------------------
+## Installing package and loading dependencies
 
 ``` r
 # devtools::install_github("trashbirdecology/bbsAssistant", dependencies = TRUE, force=FALSE)
 library(bbsAssistant)
-library(rvest)
-library(gdata)
-library(feather)
-library(here)
+library(magrittr)
 library(dplyr)
-library(tidyverse)
+library(stringr)
 ```
 
-Downloading the BBS data from USGS FTP
---------------------------------------
+## Downloading the BBS data from USGS FTP
 
 ### Define and/or create local directories
 
[…]
