+++ CONTRIBUTING.md
@@ -1,5 +1,0 @@
-# Contributing
-
-When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change to ensure the change can be implemented. Next, please submit a pull request with reference to the related Issues(s) or discussions.
-
-Please note we have a [code of conduct](CODE-OF-CONDUCT.md). You must follow it in all your interactions with the project.
