+++ README.Rmd
@@ -1,26 +1,28 @@
 ---
 title: "**bbsAssistant**: An R package for downloading and handling data and information from the North American Breeding Bird Survey."
 date: "`r Sys.Date()`"
-output: # pdf_document # when rendering PDF silence badges .svg + md_document
-  md_document:
-    variant: markdown_github
 bibliography: paper/paper.bib
+output: github_document
 ---
 
-<!-- PLEASE DO NOT EDIT README.md BY HAND. Please render README.rmd! -->
+<!-- README.md is generated from README.Rmd. Please edit that file -->
 
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/"
+)
+```
 <!-- Silence lines 12 - 13 when rendering .pdf.  -->
-[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis build status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>
+[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis build status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/> ![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)
 <!-- [![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
-![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)
-License: MIT
-
 
 # Function Descriptions and Vignettes
-For function descriptions please build the manual (`devtools::build_manual("bbsAssistant)`) and for an example build the vignette (`usethis::build_vignettes()`).
+For function descriptions please build the manual (`devtools::build_manual("bbsAssistant)`) and for an example build the vignette (`usethis::build_vignettes()`; or run `/vignettes/vignettes.Rmd`).
 
 # Contributions
-To make a contribution please submit a pr, create an [issue](/issues), or email us (jessicaleighburnett at gmail). Contributors must adhere to the [Code of Conduct](CODE_OF_CONDUCT.MD). 
+To make a contribution visit the [contributions instructions](.github/CONTRIBUTIONS.MD). Contributors must adhere to the [Code of Conduct](.github/CODE_OF_CONDUCT.MD). 
 
 # Brief Overview of bbsAssistant Functionality 
 
@@ -221,12 +223,16 @@ cred %>%
 #### Steps for obtaining argument "url" in function `get_credibility_trends`: 
 First, visit the USGS Patuxent Wildlife Research Center's [website for BBS results](https://www.mbr-pwrc.usgs.gov/) Online <https://www.mbr-pwrc.usgs.gov/>. 
 
-Next, enable the drop-down **Survey Results**, and choose **Trend Estimates (by region)** (left) and choose the desired region (e.g. Florida). Finally, copy the URL address for the page including the results and credibility measures (e.g. Florida):
+Next, enable the drop-down **Survey Results**, and choose **Trend Estimates (by region)** and choose the desired region (e.g. Florida). 
+```{r , echo=FALSE, out.width="50%"}
+knitr::include_graphics("man/figures/regcred_select_trendests_byregion.png")
+```
 
[…]
