### 2019-07-17 a28f84d
Update dictionary for spelling + auto spellcheck
Spellcheck runs on check() ignores pre-approved 'misspelled' words.
