### 2019-07-17 fa80cf4
Ignore vignette chunk runs on TRAVIS
