+++ vignettes/vignettes.Rmd
@@ -8,9 +8,11 @@ vignette: >
 ---
 
 ```{r, include = FALSE}
+TRAVIS <- !identical(tolower(Sys.getenv("TRAVIS")), "true")
 knitr::opts_chunk$set(
   collapse = TRUE,echo=TRUE, warning = FALSE, message=FALSE,
-  comment = "#>"
+  comment = "#>",
+  purl = TRAVIS # do not run chunks on TRAVIS
 )
 ```
 
