+++ NEWS.md
@@ -1,3 +1,4 @@
 # bbsAssistant 1.0.0
 
-* Added a `NEWS.md` file to track changes to the package.
+__2019.07.04__  Package v. 0.0.1 submitted for review to Journal of Open Source Software. Follow status [here](https://github.com/openjournals/joss-papers/tree/joss.01550/joss.01550).
+__2019.07.17__  Package v. 1.0.0 prepared for submission to CRAN. 
