### 2019-07-17 3b804e5
Update NEWS.md
