### 2019-07-18 #1550 @ethanwhite
I'd be happy to review this. Not sure I can guarantee 3 weeks as I have some travel coming up, but 4 should be doable.
