### 2019-07-18 017d8c6
Update vignette + ignore pdf
Ignore pdf vignette build.

Minor updates to vignette.
