### 2019-07-18 8884e0a
Revert "Update vignette + ignore pdf"
This reverts commit 017d8c62587f3d36ab2343409e3839b4203245e3.
