### 2019-07-18 69a6a53
Update vignette, ignore pdf build vign.
