### 2019-07-23 #1550 @TrashBirdEcology
@kthyng, perhaps @dombennett, @jabobsoj or @ha0ye
