### 2019-07-24 #1550 @kthyng
@dombennett are you interested and available to do a review on this JOSS submission? We ask for reviews within about 3 weeks, and you can get more information [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
