### 2019-08-01 #1550 @ha0ye
I don't think I'll be able to do 3 weeks, since I have upcoming travel all the way through Sept. 9. If you still need someone after that point, I'll be able to help out.

Some other options might be @mstrimas, @SaraVarela, or @sebpardo who each have rOpenSci packages for bird data.
