### 2019-08-02 #1550 @TrashBirdEcology
@kthyng  I am OK with the wait as the FTP server on which some of these functions relies just underwent a major overhaul without warning.  I will need to update two of the functions, which I plan to do within ~10 days.
