### 2019-08-09 77c6fe9
Add results data files + data.R
- Build warnings.
- Descriptions for data are bare bones RN
