+++ man/trend_ests_core_1966to2017.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{trend_ests_core_1966to2017}
+\alias{trend_ests_core_1966to2017}
+\title{Trend estimates for the core BBS area, 1966-2017}
+\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16971 rows and 14 columns.}
+\source{
+\url{https://www.mbr-pwrc.usgs.gov/bbs/BBS_1966-2017_core_trend_revised_v2.csv}
+}
+\usage{
+trend_ests_core_1966to2017
+}
+\description{
+More detail TBD
+}
+\keyword{datasets}
