### 2019-08-09 404400b
Added the new data in bbsAssistant in the vignette. The function to access the data needs to be added.
