### 2019-08-09 488a292
data frame for the zip files for each state. Left NA for the zip files of Mexico that have no .zip files
