+++ vignettes/vignettes.Rmd
@@ -49,6 +49,14 @@ dir.create(bbsDir)
 ### Data 
 bbsAssistant now has the latest analysis and results from the hierarchical model for population change that is available in the BBS website, which produces annual indices of abundance for a region, and subsequently estimate the trend as the ratio of the annual indices for the first and last year of interest. You can find more information on how these results were analyzed by going to the [Regional Trend Analysis website] (https://www.mbr-pwrc.usgs.gov/bbs/trend/tf15.shtml). The previous data can be found in the archives of the BBS website. To access the results and analysis in the bbsAssistant package:
 
+data(annual_index_core_1966to2017)
+glimpse(annual_index_core_1966to2017)
+
+data(annual_index_expanded_1966to2017)
+
+data(trend_ests_core_1966to2017)
+
+data(region_codes)
 
 
 
