+++ vignettes/vignettes.Rmd
@@ -32,7 +32,6 @@ library(dplyr)
 library(here)
 library(glue)
 library(rvest)
-library(magrittr)
 library(tidyr)
 ```
 
@@ -47,39 +46,47 @@ dir.create(bbsDir)
 ```
 
 ### Data 
-bbsAssistant now has the latest analysis and results from the hierarchical model for population change that is available in the BBS website, which produces annual indices of abundance for a region, and subsequently estimate the trend as the ratio of the annual indices for the first and last year of interest. You can find more information on how these results were analyzed by going to the [Regional Trend Analysis website] (https://www.mbr-pwrc.usgs.gov/bbs/trend/tf15.shtml). The previous data can be found in the archives of the BBS website. To access the results and analysis in the bbsAssistant package:
-
-data(annual_index_core_1966to2017)
-glimpse(annual_index_core_1966to2017)
-
-data(annual_index_expanded_1966to2017)
-
-data(trend_ests_core_1966to2017)
-
-data(region_codes)
-
-
-
-### Retrieve and import BBS data
-If necessary, download all or some of the BBS state-level data. Note: Downloading all the data to file takes 10-15 minutes, so only run if you have not recently downloaded the BBS data. 
-
-Let's focus on a single species and state for brevity:
-
-First, let's retrieve the regions of data that are available. The function `get_regions` retrieves the .zip filenames of all U.S. states and Canadian provinces, including their reference numbers and region codes.
-```{r, echo=TRUE, waring=FALSE, message=FALSE}
-# a. Load the regional .txt file from Patuxent's FTP server (you must be connected to the internet to perform this step)
-regions <- get_regions()
+This package contains four data objects: 
+    1. annual_index_core_1966to2017
+    1. annual_index_expanded_1966to2017
+    1. trend_ests_core_1966to2017
+    1. region_codes
+    
+Each data can be loaded into the environment using the function, `utils::data`. E.g., data("region_codes"). The data `region_codes` contains each country/state/province and the filename for the associated compressed (.zip) folder in the FTP server containing all BBS data. 
+```{r}
+data("region_codes")
+glimpse(region_codes)
+# unique(region_codes$CountryNum) # country numbers
+# unique(region_codes$State) # unique US and Mexican states, and Canadian provinces
 ```
 
-Let's restrict our data download to **Florida** data:
+Data objcects `annual_index_core_1966to2017`, `annual_index_expanded_1966to2017`, and `trend_ests_core_1966to2017` contain the results for each Region x Year x Species (AOU) from the hierarhical population trend models. Model results include annual population trend indices for each region/species combination. Details of the modelsa and results can be found on the BBS's [Regional Trend Analysis webpage] (https://www.mbr-pwrc.usgs.gov/bbs/trend/tf15.shtml). 
+
+```{r}
+data("annual_index_core_1966to2017")
+glimpse(annual_index_core_1966to2017)
+
+# data("annual_index_expanded_1966to2017")
+# data("trend_ests_core_1966to2017")
+```
+
+### Retrieve and import BBS data
+
+Easily download all or select regions of the BBS state-level count data. Please note that downloading **all** BBS state-level data will take approximately 10-15 minutes, so only run this if necessary. For brevity, let's download only the state Florida data. 
+
+First, load the region filenames into memory (data) filenames of all U.S. states and Canadian provinces, including their reference numbers and region codes.
+```{r, echo=TRUE, waring=FALSE, message=FALSE}
+data("region_codes")
+```
+
[…]

[…]
