### 2019-08-16 d2bfe7d
Remove fxn: get_regions
+ minor update to get_credibility_trends (grammar only)
