+++ NAMESPACE
@@ -5,7 +5,6 @@ export(export_bbsFeathers)
 export(get_analysis_results)
 export(get_bbsData)
 export(get_credibility_trends)
-export(get_regions)
 export(get_routeInfo)
 export(get_speciesList)
 export(get_unzip)
