### 2019-08-16 57f7378
Revert "Remove fxn: get_regions"
This reverts commit d2bfe7df1991a1cad4c602f65003d08daa2f96a9.
