+++ R/get_credibility_trends.R
@@ -1,6 +1,6 @@
 #' @title Get data credibility and trends from the BBS analysis results
 #' @description The BBS provides regional and state-wide species population trend estimates. They also include a 'data credibility' rating (low, medium, or high deficiencies) for each species-region combination. This function uses web scraping to capture population trends and their credibility ratings.
-#' @param url Web address of the region or state for which species' population trends and credibility ratings are to be scraped. Default example is the state 'Kansas, US'.
+#' @param url Web address of the region or state for which species' population trends and credibility ratings are to be scraped. Default example is 'Kansas,USA'.
 #' @export get_credibility_trends
 #' @importFrom magrittr %>%
 
