### 2019-08-16 353b0bf
Remove get_regions.R
+ minor updates to documentation.
