+++ man/get_regions.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_regions.R
-\name{get_regions}
-\alias{get_regions}
-\title{Get the list of available data regions from the BBS website}
-\usage{
-
-  get_regions(source = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/",
-  bbsDir = NULL)
-}
-\arguments{
-\item{source}{The website or location of the BBS data files.}
-
-\item{bbsDir}{The location of the local folder containing the bbs raw data. If this folder DNE, a new folder will be created in the working directory.}
-}
-\description{
-Read in list of regions (State/Prov/TerrName), from RegionCodes.txt, and then extract list of where the 10-stop data is kept. This function was adapted from the function \href{https://github.com/oharar/rBBS/blob/master/R/GetRegions.R}{rBBS::GetRegions}.
-}
