+++ man/get_routeInfo.Rd
@@ -28,7 +28,7 @@ This function downloads information about route location from the BBS FTP server
 \examples{
 # download BBS route data.
 \dontrun{
-RouteInfo <- get_routeInfo()
+routeInfo <- get_routeInfo()
 }
 
 }
