+++ NAMESPACE
@@ -1,5 +1,16 @@
 # Generated by roxygen2: do not edit by hand
 
+export("%>%")
+export(export_bbsFeathers)
+export(get_analysis_results)
+export(get_bbsData)
+export(get_credibility_trends)
+export(get_routeInfo)
+export(get_speciesList)
+export(get_unzip)
+export(import_bbsData)
+export(import_bbsFeathers)
+export(subset_speciesList)
 importFrom(magrittr,"%>%")
 importFrom(stats,family)
 importFrom(utils,download.file)
