### 2019-08-16 2024a4f
Update vignette
