+++ R/data.R
@@ -26,7 +26,7 @@
 
 
 # annual_index_expanded_1966to2017 ----------------------------------------
-#' Annual population trend indices for species/regions in the exanded BBS region, 1966-2017
+#' Annual population trend indices for species/regions in the expanded BBS region, 1966-2017
 #'
 #' Description TBD
 #'
@@ -35,9 +35,9 @@
 #'   \item{AOU}{Five digit species identification number according to the bird species checklist of the American Ornithologist Union (AOU)}
 #'   \item{Region}{Two letter abreviation for a US state or region}
 #'   \item{Year}{Four digit year of survey, from 1966-2017}
-#'   \item{Index}{}
-#'   \item{2.5% CI}{}
-#'   \item{97.5% CI}{}
+#'   \item{Index}{Annual index of species/region abundance estimated using a hierarchical model}
+#'   \item{2.5% CI}{Lower 95% confidence interval (2.5%) of the annual index of abundance}
+#'   \item{97.5% CI}{Upper 95% confidence interval (97.5%) of the annual index of abundance}
 #' }
 #' @source \url{https://www.mbr-pwrc.usgs.gov/bbs/inde_1993-2017_expanded.csv}
 "annual_index_expanded_1966to2017"
@@ -53,9 +53,9 @@
 #'   \item{AOU}{Five digit species identification number according to the bird species checklist of the American Ornithologist Union (AOU)}
 #'   \item{Region}{Two letter abreviation for a US state or region}
 #'   \item{Year}{Four digit year of survey, from 1966-2017}
-#'   \item{Index}{}
-#'   \item{2.5% CI}{}
-#'   \item{97.5% CI}{}
+#'   \item{Index}{Annual index of species/region abundance estimated using a hierarchical model}
+#'   \item{2.5% CI}{Lower 95% confidence interval (2.5%) of the annual index of abundance}
+#'   \item{97.5% CI}{Upper 95% confidence interval (97.5%) of the annual index of abundance}
 #' }
 #' @source \url{https://www.mbr-pwrc.usgs.gov/bbs/inde_1966-2017_core_v2.csv}
 "annual_index_core_1966to2017"
