### 2019-08-20 664f696
Final descritpions of data.R
