### 2019-08-21 865a24b
Fix typo (Wisconsin) in region_codes df
Fixed bug #55
