### 2019-08-27 bae1c1b
Fix a couple of typos
