### 2019-08-27 9ae3fa3
Make function name consistent in get_bbsData help file
