### 2019-09-10 8efe4f8
Update the vignette and rebuild the vignette for .io
Major edits to the vignette to reflect dataset additions.

Used pkgdown::build_articles() to update the webiste vignette.
