+++ vignettes/vignettes.Rmd
@@ -6,7 +6,6 @@ vignette: >
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
-
 ```{r, include = FALSE}
 TRAVIS <- !identical(tolower(Sys.getenv("TRAVIS")), "true")
 knitr::opts_chunk$set(
@@ -15,7 +14,6 @@ knitr::opts_chunk$set(
   purl = TRAVIS # do not run chunks on TRAVIS
 )
 ```
-
 # Brief Overview of bbsAssistant Functionality 
 ## Installing package and loading dependencies
 ```{r loadDeps, echo=TRUE, warning=FALSE, message=FALSE}
@@ -53,7 +51,7 @@ This package contains four data objects:
     1. region_codes
     
 Each data can be loaded into the environment using the function, `utils::data`. E.g., data("region_codes"). The data `region_codes` contains each country/state/province and the filename for the associated compressed (.zip) folder in the FTP server containing all BBS data. 
-```{r}
+```{r region_codes}
 data("region_codes")
 glimpse(region_codes)
 # unique(region_codes$CountryNum) # country numbers
@@ -62,7 +60,7 @@ glimpse(region_codes)
 
 Data objcects `annual_index_core_1966to2017`, `annual_index_expanded_1966to2017`, and `trend_ests_core_1966to2017` contain the results for each Region x Year x Species (AOU) from the hierarhical population trend models. Model results include annual population trend indices for each region/species combination. Details of the modelsa and results can be found on the BBS's [Regional Trend Analysis webpage] (https://www.mbr-pwrc.usgs.gov/bbs/trend/tf15.shtml). 
 
-```{r}
+```{r annualindex}
 data("annual_index_core_1966to2017")
 glimpse(annual_index_core_1966to2017)
 
@@ -75,38 +73,38 @@ glimpse(annual_index_core_1966to2017)
 Easily download all or select regions of the BBS state-level count data. Please note that downloading **all** BBS state-level data will take approximately 10-15 minutes, so only run this if necessary. For brevity, let's download only the state Florida data. 
 
 First, load the region filenames into memory (data) filenames of all U.S. states and Canadian provinces, including their reference numbers and region codes.
-```{r, echo=TRUE, waring=FALSE, message=FALSE}
+```{r, echo=TRUE, warning=FALSE, message=FALSE}
 data("region_codes")
+glimpse(region_codes)
 ```
 
 Find the .zip file name associated with the state, Florida:
-```{r, echo=TRUE}
+```{r statenames, echo=TRUE}
 regionFileName <- region_codes$zip_states
 (regionFileName.use <- regionFileName[which(stringr::str_detect(regionFileName, "Flori")==TRUE)])
 ```
 
 Once we have one or more region filenames, we can use function `get_bbsData` to download the .zip file to a temporary folder (unless otherwise specified), and *import* the temp file to R object. The R object, flBBS, contains the raw BBS count data.
-```{r, echo=TRUE}
-require(bbsAssistant)
+```{r getbbsdata, echo=TRUE}
 flBBS <- get_bbsData(file=regionFileName.use)
 ```
 
 Next, we can download the BBS route-level geographic information and metadata, and append this to the original data.
-```{r, echo=TRUE}
+```{r getroutes, echo=TRUE}
 routes <- get_routeInfo() # retrieve route-level data
 flBBS <- dplyr::left_join(flBBS, routes) # merge route-level data to bird count data
 glimpse(flBBS %>% dplyr::select(aou, year, route, statenum, countrynum, stoptotal, latitude, longitude))
 ```
 
[…]
