+++ R/data.R
@@ -63,9 +63,9 @@
 
 
 # region_codes ----------------------------------------
-#' Annual population trend indices for species/regions in the core BBS region, 1966-2017
+#' Contains country and state numbers, and the associated zip filenames.#' 
 #'
-#' Description TBD
+#' 
 #'
 #' @format Describe the format of df:
 #' \describe{
@@ -74,8 +74,8 @@
 #'   \item{State}{The proper name of the (US/MEX) state or (CAN) province}
 #'   \item{zip_states}{The filename of the .zip compressed BBS count data located on the FTP server. These files do not exist for Mexican states, so == NA}
 #' }
-#' @source \url{https://www.mbr-pwrc.usgs.gov/bbs/inde_1966-2017_core_v2.csv}
-"annual_index_core_1966to2017"
+#' @source \url{ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/RegionCodes.txt}
+"region_codes"
 
 
 
