### 2019-09-10 032d888
Edit the JOSS paper
