### 2019-09-12 28c4864
Fix typos per Gabby
