### 2019-09-16 #1550 @kthyng
Great! We're ready to continue then! Maybe we'll have better luck with reviewers' schedules now too...

@ethanwhite and @jhollist Would you be able to review this submission now that it is ready to go? As before, we'd ask for reviews in about the next 3 weeks or so.
