### 2019-09-17 #1550 @ha0ye
I think I can find the time over the next 3-4 weeks.

As a heads up though, I'm in the same lab as @ethanwhite, and we have our own project that includes a component that processes data (including BBS). Also, I know @TrashBirdEcology personally, and we have collaborated on work in the past.

I don't think those are substantial obstacles to my providing a constructive review; the overlap with @ethanwhite might be the most challenging in terms of finding a diverse set of reviewers, but I can understand the difficulties in wrangling sufficient reviewers to begin with!
