### 2019-09-17 #1550 @kthyng
Thanks @ha0ye, makes sense that just you or @ethanwhite would be reviewers. Hopefully we can hear about @ethanwhite about his availability.

@mstrimas, @SaraVarela, or @sebpardo are you interested and available to do a review on this JOSS submission? We ask for reviews within about 3 weeks, and you can get more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .
