### 2019-09-20 #1550 @TrashBirdEcology
@kthyng I have posted on rstats twitter to see if we can't find someone (who isn't yet a JOSS reviewer)!
