### 2019-09-23 #1550 @kthyng
@mstrimas We would prefer a code review actually. Maybe you can review for JOSS another time. Thanks!
