### 2019-09-26 #1768 @whedon
[REVIEW]: bbsAssistant: An R package for downloading and handling data and information from the North American Breeding Bird Survey.
**Submitting author:** @trashbirdecology (<a href="http://orcid.org/0000-0002-0896-5099">Jessica Burnett</a>)
**Repository:** <a href="https://github.com/TrashBirdEcology/bbsAssistant/" target ="_blank">https://github.com/TrashBirdEcology/bbsAssistant/</a>
**Version:** v1.0
**Reviewer:** @ethanwhite , @jsta
**Archive:** <a href="https://doi.org/10.5066/P93W0EAW" target="_blank">10.5066/P93W0EAW</a>
