### 2019-10-16 24425a0
spellcheck fixes
