+++ paper/paper.md
@@ -31,7 +31,7 @@ affiliations:
 This package contains functions for downloading and munging data from the North American Breeding Bird Survey (BBS) FTP server [@pardieck2018north; @sauer2017first]. Although the FTP server provides a public interface for retrieving data and analysis results, this package consolidates the efforts of the data user by automating downloading and decompression of .zip data files, downloading route-level information, and saving them as .feather files for speedy import from disk. The data subsetting features of this package also allow the user to readily import and save to file only the data necessary for their purposes. Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
 
 # Retrieving and Munging North American Breeding Bird Survey Data 
-The BBS team uses hierarhical modelling techniques to generate population trend estimates [@sauer2017north] at various spatial scales [see the BBS results webpage](https://www.mbr-pwrc.usgs.gov/). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package retrieves the raw count BBS data from the U.S. Geological Survey's FTP server using function `get_bbsData()`. This function requires an on-line connection, and contains arguments for subsetting the data by species identity and/or region. The data are first downloaded into a temporary folder and then uploaded into the R workspace as a single data frame object. The function 
+The BBS team uses hierarchical modelling techniques to generate population trend estimates [@sauer2017north] at various spatial scales [see the BBS results webpage](https://www.mbr-pwrc.usgs.gov/). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package retrieves the raw count BBS data from the U.S. Geological Survey's FTP server using function `get_bbsData()`. This function requires an on-line connection, and contains arguments for subsetting the data by species identity and/or region. The data are first downloaded into a temporary folder and then uploaded into the R workspace as a single data frame object. The function 
 `export_bbsFeathers()` saves the data frame to a specified directory as .feather files, allowing for long-term local storage of the raw data. Importing the .feathers is expedited using the function `import_bbsFeathers()`, and provides arguments for further subsetting the data prior to loading it into the environment. Used in conjunction with the species list (`get_speciesList()`), post-hoc subsetting is simplified by the function `subset_speciesList()`. 
 
 # Retrieving Population Trend Model Results from the BBS website
