### 2019-10-16 TrashBirdEcology/bbsAssistant#63 @jsta
JOSS Review Issues
I figured I would start an issue thread here for the JOSS Review https://github.com/openjournals/joss-reviews/issues/1768

My first set of comments (1) are more or less "must-haves" to meet the JOSS requirements while the second set (2) are "nice-to-have" suggestions.

## 1_JOSS

- [x] `paper.md` should probably have an explicit citation for the `rBBS` package.

- [x] The `testthat` tests in `test-get_regions.R` fails for me with error:

> Error in get_regions() : could not find function "get_regions"

## 2_Nice-to-haves

### User friendliness

The stated goal of the package is to make working with bbs data more user friendly by consolidating "the efforts of the data user by automating downloading and decompression of .zip data files, downloading route-level information, and saving them as .feather files for speedy import from disk."

As a new user, there were some areas that I thought could be even more user friendly:

- [x] The biggest was the manual download > export > import process. Current operation is download to temporary file then manually export to "permanent" disk then later import this data. What if this was revised so that a user could optionally pass a permenent location to `get_bbsData`, this location would checked for an existing file (to avoid re-downloading existing files) and automatically exported?

- [ ] I wish the docs included some maps. For instance, as a new user I didn't know the difference between `Region=="S05"` and `Region.Code=="FLA"`

- [x] I felt like some of the calls to the `stringr` package in the vignette were not needed. Maybe people who don't know `stringr` would have an easier time if these were replaced with base R code? Maybe the `region_code` object could be cleaned up so that someone could run `region_codes[region_codes$State == "Florida", "zip_states"]`?

- [x] The vignette/package uses the feather package AND variables called `feather`. This could be confusing.

- [ ] Maybe the vignette could use a more descriptive title rather than "vignettes"?

### Misc

- [x] I had trouble understanding the `subset_speciesList` function. Why does it remove the specified family rather than include it like `dplyr::filter`? What does this function offer
beyond `dplyr::filter`?

- [x] Why are the first and second figures in the vignette different? Seems like they are both plotting `AOU.Number=='s06882'` and `Region.Code=="S05"` by year.

### pkgdown

- [x] Seems like the `pkgdown` site is out of date. The current landing page doesn't match the README.

- [x] Vignette images on the pkgdown site are broken for me...

- [x] The vignette chunk `Subsetting the BBS count data by AOU number` prints A LOT of output. Maybe it would be good to pass the output through `glimpse` like the other chunk outputs?
