### 2019-10-17 TrashBirdEcology/bbsAssistant#63 @kthyng
@TrashBirdEcology and @jsta Great to deal with issues in separate issues from the overall review like this. Also might make sense to have multiple separate issues for some of these items, depending on how separate and major they are. I'll let you organize though.
