### 2019-10-17 TrashBirdEcology/bbsAssistant#63 @trashbirdecology
@jsta Thank you for the valuable feedback! Will address as soon as possible. Keeping open until all issues resolved or re-assigned.
