### 2019-11-01 #1768 @kthyng
@ethanwhite Just a friendly ping here.
