### 2019-11-05 ee844d4
Fix typo in vignette
