### 2019-11-05 #1768 @ethanwhite
I've completed my checklist level review. Below are sections related to fixes for boxes that I haven't checked yet:

## Automated tests

* Two tests are currently failing in my local build for `test-get_regions.R`, but they seem to be passing on Travis so maybe I have a missing dependency or something. Ubuntu 18.04, R 3.6.1.

## Community Guidelines

* Contributing and CoC links were broken. Fixed in https://github.com/TrashBirdEcology/bbsAssistant/pull/65
* No clear description of how to "Report issues or problems with software" or "Seek support". There is some language related to reporting issues in
  `CONTRIBUTING.md`.

## State of the field

There are two additional packages that contain some of the functionality described here:

1. `rBBS` - This  package is acknowledged as a source of some code, but no comparison is made to it in the paper.
2. [`rdataretriever`](https://github.com/ropensci/rdataretriever) (software by my group) - which downloads BBS and loads it into R. This package works on the entire integrated BBS dataset, does not support data manipulation, and does not work with the BBS analyses, so there are definitely meaningful differences from this package and therefore no conflict with publishing this software paper. However, it does have some meaningful overlap and so may be worth mentioning here. It is potentially a useful alternative for users looking to work with a full integrated version of the BBS data.

## References

* The `rBBS` package should be cited.
* If `rdataretriever` is added then it and the associated `retriever` package (https://doi.org/10.21105/joss.00451)

I have some additional suggestions related to the functionality of the package and the API, but those are out of scope for the JOSS review checklist so I will open them as issues over in the `bbsAssistant` repository. (I will try to remember to tag this issue in case there is interest in considering them as part of this review).
