### 2019-11-05 TrashBirdEcology/bbsAssistant#66 @ethanwhite
Confusing API for `subset_speciesList`
This is recommendation based on my experience while reviewing this package for JOSS (https://github.com/openjournals/joss-reviews/issues/1768).

I was confused by the API for `subset_speciesList`. My reading of the vignette was that passing `aou.ind` would subset the data to only the provided AOU. This matched my intuition for how a function named `subset_speciesList` would work. However, as documented in the function documentation the behavior is actually the opposite, with the list of species provided being removed and all other species kept.

**Describe the solution you'd like**

I would consider either renaming this function to something like `remove_species` or changing the argument names to something like `drop_species`. This would make the i

Clarifying the vignette to make the behavior clearer would be another option, but I think intuitive naming would go a long way to preventing user confusion.
