### 2019-11-05 TrashBirdEcology/bbsAssistant#67 @ethanwhite
Focus top-level API / vignette on key user behavior
This is recommendation based on my experience while reviewing this package for JOSS (https://github.com/openjournals/joss-reviews/issues/1768).

As a new user of the software I found the API as introduced in the vignette to be a bit complicated. The order of introduction is currently: 1) Create a data directory for storing data; 2) Download and load data into R (this data doesn't end up in the data directory); 3) Export the data into a local file (in a not widely used format; I love `feather` but it won't be familiar to many users); 4) Import the data (which requires doing some filename matching and manipulation). This felt like a lot of stuff if my goal is to quickly and easily work the the BBS data.

**Describe the solution you'd like**

My core use case for a package like this would be "Get BBS data into R and start working with it. Ideally without repeatedly downloading the data if I already have it". With that use case in mind I would consider making `get_bbsData` handle all of this for the user, with some optional arguments to control behavior. So, the default behavior for `get_bbsData` would be to:

1. Check if the requested data exists locally
2. If it doesn't then download the data and store it locally in a permanent (not temporary) location.
3. Load the data into R.

I'd have an argument to allow setting the data directory (where files are checked for and downloaded to) and set a default for this directory to either the working directory (in a clearly named subdirectory) or a `.bbsAssistant` directory in the users home directory.

Given that most of the intended usage appears to focus on single BBS files, which are relatively small, another option would be to de-emphasize the `feather` storing and loading functionality in the introductory vignette.
