### 2019-11-05 TrashBirdEcology/bbsAssistant#63 @ethanwhite
In the spirit of providing an independent review I held off on looking at this issue until I was done. Now that I'm finished and taking a look it looks like @jsta & I highlighted a number of similar things in the UI, so :raised_hands: for consistent reviews!
