### 2019-11-05 TrashBirdEcology/bbsAssistant#63 @trashbirdecology
Thanks, Ethan!

Jessica L. Burnett

GitHub <http://github.com/trashbirdecology>
CV/Resume
<https://github.com/TrashBirdEcology/cv/blob/master/burnett_extendedCV.pdf>
ORC ID: 0000-0002-0896-5099 <https://orcid.org/0000-0002-0896-5099>

On Tue, Nov 5, 2019 at 12:43 PM Ethan White <notifications@github.com>
wrote:

> In the spirit of providing an independent review I held off on looking at
> this issue until I was done. Now that I'm finished and taking a look it
> looks like @jsta <https://github.com/jsta> & I highlighted a number of
> similar things in the UI, so 🙌 for consistent reviews!
>
