+++ README.Rmd
@@ -34,7 +34,7 @@ For function descriptions please build the manual (`devtools::build_manual("bbsA
 To make a contribution visit the [contributions instructions](.github/CONTRIBUTING.md). Contributors must adhere to the [Code of Conduct](.github/CODE_OF_CONDUCT.md). 
 
 ## Acknowledgments
-We thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. Some functions in this package were adapted from the [rBBS](github.com/oharar/rbbs) package and are mentioned in function source code as applicable.
+We thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. Some functions in this package were adapted from the [rBBS](https://github.com/oharar/rbbs) package and are mentioned in function source code as applicable.
 
 
 ## References
