### 2019-11-06 28b141a
Update links to community files
It looks like CRAN rejects packages with relative links to the `.github` folder lately because they result in broken links when they render the README on their servers.
