### 2019-11-14 #1768 @kthyng
Hi @TrashBirdEcology! Just want to make sure you saw that @ethanwhite has some recommendations for you for moving forward.
