### 2019-11-14 #1768 @TrashBirdEcology
Indeed! Have been travelling, need to hop on this! Thanks to both of you!

Jessica L. Burnett

GitHub <http://github.com/trashbirdecology>
CV/Resume
<https://github.com/TrashBirdEcology/cv/blob/master/burnett_extendedCV.pdf>
ORC ID: 0000-0002-0896-5099 <https://orcid.org/0000-0002-0896-5099>

On Thu, Nov 14, 2019 at 11:42 AM Kristen Thyng <notifications@github.com>
wrote:

> Hi @TrashBirdEcology <https://github.com/TrashBirdEcology>! Just want to
> make sure you saw that @ethanwhite <https://github.com/ethanwhite> has
> some recommendations for you for moving forward.
>
