### 2019-11-14 TrashBirdEcology/bbsAssistant#66 @trashbirdecology
Agreed. Will update to include remove_spp.. Not sure what I was thinking..
