### 2019-11-15 TrashBirdEcology/bbsAssistant#67 @trashbirdecology
Thanks for the comments. I agree with the updates to retrieving, storing and importing the raw data.

@ethanwhite I am curious your thoughts on feather. The only reason I chose this was to reduce filesize and import times for users who are working with a large portion of the data. I suppose I could just create file extension options while defaulting to a TSV, TXT, etc.
