### 2019-11-15 TrashBirdEcology/bbsAssistant#67 @ethanwhite
I like feather in concept, but since adoption in the ecology world is somewhat limited I think it's an open question as to whether it's a good choice here. Given the size of the data I think the biggest benefit is using the package to educate ecologists about the format. Since BBS is <1GB in total I thing csv/tsv are both reasonable options that allow the data generated to be used cross-language. If you want the storage space read-time benefits of binary then since the usage is really in R I might just use the R data format. That said, as long as you're abstracting this from the user then I don't think it really matters. They just care about `get_bbsData` given them data in R, what the back end storage is won't matter to them - which means that any of these choices is perfectly reasonable.
