### 2019-11-15 TrashBirdEcology/bbsAssistant#67 @trashbirdecology
True that
