### 2019-11-19 59a8b36
Major updates. Addressing major suggestions from JOSS reviewers.
