### 2019-11-20 1800523
major edits
