### 2019-11-21 TrashBirdEcology/bbsAssistant#68 @trashbirdecology
Resolved in function `download_bbs`.
