### 2019-11-21 18c7485
update download and import functions and vignette.
