+++ R/get_routeInfo.R
@@ -1,53 +1,53 @@
-#' @title Download route information from USGS server
-#' @description This function downloads information about route location from the BBS FTP server. This function was adapted from **oharar/rBBS** package.
-#' @param routesDir Location of the routes.zip folder Should be in DatFiles folder (default).
-#' @param RouteTypeID One or more numbers indicating route substrate (1=roadside;2=water;3=off-road; Default = 1, roadside only).
-#' @param Stratum A vector of BBS physiographic stratum codes by which to filter the routes.
-#' @param BCR A vector of Bird Conservation Region codes where by which to filter the routes.
-#' @return If download successful, a dataframe with the results.
-#' @importFrom magrittr %>%
-#' @export get_routeInfo
-#' @examples
-#' # download BBS route data.
-#' \dontrun{
-#' routeInfo <- get_routeInfo()
+#' #' @title Download route information from USGS server
+#' #' @description This function downloads information about route location from the BBS FTP server. This function was adapted from **oharar/rBBS** package.
+#' #' @param routesDir Location of the routes.zip folder Should be in DatFiles folder (default).
+#' #' @param RouteTypeID One or more numbers indicating route substrate (1=roadside;2=water;3=off-road; Default = 1, roadside only).
+#' #' @param Stratum A vector of BBS physiographic stratum codes by which to filter the routes.
+#' #' @param BCR A vector of Bird Conservation Region codes where by which to filter the routes.
+#' #' @return If download successful, a dataframe with the results.
+#' #' @importFrom magrittr %>%
+#' #' @export get_routeInfo
+#' #' @examples
+#' #' # download BBS route data.
+#' #' \dontrun{
+#' #' routeInfo <- get_routeInfo()
+#' #' }
+#' #'
+#' 
+#' get_routeInfo <- function(
+#'                      routesDir =  "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/",
+#'                          RouteTypeID = 1,
+#'                          # one or more of c(1,2,3)
+#'                          Stratum = NULL,
+#'                          BCR = NULL) {
+#'     # Unzip from FTP server and store as an R object
+#'     routeDat <-
+#'         get_unzip(
+#'             ZipName = paste0(routesDir, routes.zip),
+#'             FileName = gsub("^Fifty", "fifty", gsub("zip", "csv", routes.zip))
+#'         )
+#'     
+#'     # str(routeDat )
+#'     # Force column names to lowercase
+#'     names(routeDat) <- tolower(names(routeDat))
+#'     
+#'     # Filter the routes if specified
+#'     
+#'         if (!is.null(Stratum)) {
+#'             routeDat <- routeDat %>%
+#'                 dplyr::filter(stratum %in% Stratum)
+#'         }
+#'         
+#'         if (!is.null(BCR)) {
+#'             routeDat <- routeDat %>%
+#'                 dplyr::filter(BCR %in% BCR)
+#'         }
+#'         
+#'         
+#'         if (!is.null(RouteTypeID)) {
+#'             routeDat <- routeDat %>%
+#'                 dplyr::filter(RouteTypeID %in% RouteTypeID)
+#'         }
+#' 
+#'     return(routeDat)
 #' }
-#'
-
-get_routeInfo <- function(
-                     routesDir =  "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/",
-                         RouteTypeID = 1,
-                         # one or more of c(1,2,3)
-                         Stratum = NULL,
-                         BCR = NULL) {
-    # Unzip from FTP server and store as an R object
-    routeDat <-
-        get_unzip(
[…]
