### 2019-11-22 2141d23
Major function updates...
