+++ R/get_species.R
@@ -1,0 +1,100 @@
+#' @title Download and import BBS, NACC, and AOU species list
+#' @description Read in list of species names, alpha codes, and taxonomic information from various sources (BBS, AOU/AOS, IBP, NACC).
+#' For more information on species alpha codes please visit https://www.birdpop.org/pages/birdSpeciesCodes.php
+#' For more information on AOU/AOS/NACC species checklists please visit http://checklist.aou.org/
+#'    Portions of this function (lines for importing and munging the BBS species list, specieslist.txt) were borrowed from the function [`rBBS::getSpNames()`](https://github.com/oharar/rBBS/blob/master/R/GetSpNames.R). This package is not actively maintained, therefore we export it via `bbsAssistant`.
+#' @param bbs.spp.url URL for SpeciesList.txt file, retrieved via FTP.
+#' @param nacc.spp.url URL for the NACC species checklist, which contains taxonomic sorting information no availabble in BBS SpeciesList.
+#' @param aou.alpha URL for the 4- and 6-letter alpha codes associated with latin names (source: IBP).
+#' @param data.dir Location of where to save the compressed and decompressed species lists and taxonomic files. Defaults to ~/raw-data
+#' @export get_species
+#' @return A list (`species.lists`) comprising the BBS species list (`bbs.spp`) and the [NACC species checklist](http://checklist.aou.org/taxa.csv?type=charset%3Dutf-8%3Bsubspecies%3Dno%3B) (`nacc.spp`), and the AOU alpha checklist (`aou.alpha`). 
+
+get_species <- function(bbs.spp.url = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/SpeciesList.txt",
+                        nacc.spp.url = "http://checklist.aou.org/taxa.csv?type=charset%3Dutf-8%3Bsubspecies%3Dno%3B",
+                        aou.alpha.url = "https://www.birdpop.org/docs/misc/IBPAOU.zip",
+                        data.dir = paste0(getwd(), "/raw-data/")) {
+    # Retrieve and munge the BBS specieslist.txt
+    temp <- list.files(data.dir, pattern = "SpeciesList.txt")
+    if (length(temp) == 0) {
+        # if it does not exist then download, munge, and write locally to avoid having to create internet connection.
+        All <-
+            scan(
+                bbs.spp.url,
+                what = "character",
+                sep = "\n",
+                encoding = "latin1"
+            )
+        Delimiter <- grep("^-", All)
+        
+        ColNames <-
+            strsplit(All[Delimiter - 1], split = '[[:blank:]]+')[[1]]
+        Widths <-
+            nchar(strsplit(All[Delimiter], split = '[[:blank:]]+')[[1]])
+        
+        Lines <- sapply(All[-(1:Delimiter)], function(str, w) {
+            trimws(substring(str, c(1, cumsum(w[-length(w)])), cumsum(w)))
+        }, w = Widths + 1)
+        colnames(Lines) <- NULL
+        rownames(Lines) <- ColNames
+        
+        bbs.spp <- as.data.frame(t(Lines), stringsAsFactors = FALSE)
+        bbs.spp$Seq <- as.numeric(bbs.spp$Seq)
+        bbs.spp$AOU <- as.numeric(bbs.spp$AOU)
+        names(bbs.spp) <-
+            c(
+                "seq",
+                "aou",
+                "common_eng",
+                "common_fr",
+                "latin",
+                "order",
+                "family",
+                "genus",
+                "species"
+            )
+        
+        write.table(x = bbs.spp,
[…]
