+++ paper/paper.bib
@@ -4,7 +4,6 @@
   year={2017}
 }
 
-
 @article{sauer2017first,
   title={The first 50 years of the North American Breeding Bird Survey},
   author={Sauer, John R and Pardieck, Keith L and Ziolkowski Jr, David J and Smith, Adam C and Hudson, Marie-Anne R and Rodriguez, Vicente and Berlanga, Humberto and Niven, Daniel K and Link, William A},
