### 2019-11-22 fdfbb1d
Major updates to vignette to address changes.
