+++ vignettes/quick-overview.Rmd
@@ -5,6 +5,7 @@ vignette: >
   %\VignetteIndexEntry{vignettes}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+bibliography: paper/paper.bib
 ---
 ```{r, include = FALSE}
 TRAVIS <- !identical(tolower(Sys.getenv("TRAVIS")), "true")
@@ -33,12 +34,12 @@ library(tidyr)
 ```
 
 # Brief Overview of bbsAssistant Functionality 
-This vignette gives a quick overview of the primary functions of the bbsAssistant package:
+This vignette gives a quick overview of the primary functions of the `bbsAssistant` package:
 1. Retrieving, downloading, and manipulating the annually-updated BBS data from the USGS servers via FTP.
 1. Loading and exploring the data stored in the package (updated annually by maintainer or as-needed)
 1. Visualizing BBS data and official BBS model results 
 
-## A: Retrieving, Downloading, and Manipulating route-level BBS data
+## I: Retrieving, Downloading, and Manipulating route-level BBS data
 This package was creating for expediting the BBS data retrieval process, which currently comprises multiple .zip files retrievable individually via FTP server, and is particularly useful for to those wishing to quickly download and manipulate route-level data from  _multiple regions/states_. 
 
 ### Download the BBS data from USGS
@@ -58,187 +59,116 @@ download_bbs(country.names="USA", data.dir = bbsDir)
 Now, let's import the data into R by first decompressing each .zip files of interest from subdirectory `~/raw-data`, and loading each into a list element. Finally, we will condense all into a single data frame. Let's import Florida and Nebraska only:
 ```{r import_bbs}
 bbs.df <- import_bbs(data.dir = bbsDir,state.names=c("FlOrIda", "NEBRASKa")) # capitalization irrevlevant.
-str(bbs.df)
+# This data frame differs slightly from BBS raw data in that we have appended State Names (`State`) to the df
+bbs.df$State %>% unique()
 ```
 
+### Download and route-level information
+Retrieve the route-level information only. This data contains basic information about route location (lat, lon, state, country), activity, BCR location, stratum (BBS), and route name. 
 ```{r routes}
 routes <- get_routes(data.dir=bbsDir)
 ```
 
+Retrieve the local conditions associated with each route-year iteration. 
 ```{r}
-
+conditions <- get_conditions(data.dir=bbsDir)
 ```
 
-# END RUN
-
-**NOTE**: If the directory exists, it will not overwrite files. If the bbs data already exists inside bbsDir, then we will create a logical to NOT download the data (see below). If you wish to download more, or overwrite existing data, please specify downloadBBSData=TRUE or remove .zip files from **/bbsData/**.
-
-
-## Data Stored in the Package
-This package contains four data objects: 
-    1. annual_index_core_1966to2017
-    1. annual_index_expanded_1966to2017
-    1. trend_ests_core_1966to2017
-    1. region_codes
-    
-Each data can be loaded into the environment using the function, `utils::data`. E.g., data("region_codes"). The data `region_codes` contains each country/state/province and the filename for the associated compressed (.zip) folder in the FTP server containing all BBS data. 
-```{r region_codes}
-data("region_codes")
-glimpse(region_codes)
-# unique(region_codes$CountryNum) # country numbers
-# unique(region_codes$State) # unique US and Mexican states, and Canadian provinces
+### Append taxonomic information and species indices
[…]
