+++ paper/paper.bib
@@ -30,9 +30,3 @@
   journal={Patuxent Wildlife Research Center, Laurel, Maryland},
   year={1997}
 }
-
-@software{rbbs,
-  author = {O'Hara, Bob},
-  title = {rBBS: R package to import USGS' BBS data},
-  url = {https://www.github.com/oharar/rbbs}
-}
