+++ README.Rmd
@@ -22,9 +22,6 @@ Install the development version of this package using devtools and load the pack
 devtools::install_github("trashbirdecology/bbsAssistant", 
                          dependencies = TRUE, force=FALSE)
 library(bbsAssistant)
-library(magrittr)
-library(dplyr)
-library(stringr)
 ```
 
 ## Function Descriptions and Vignettes
@@ -34,6 +31,6 @@ For function descriptions please build the manual (`devtools::build_manual("bbsA
 To make a contribution visit the [contributions instructions](https://trashbirdecology.github.io/bbsAssistant/CONTRIBUTING.html). Contributors must adhere to the [Code of Conduct](https://trashbirdecology.github.io/bbsAssistant/CODE_OF_CONDUCT.html). 
 
 ## Acknowledgments
-We thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. Some functions in this package were adapted from the [rBBS](https://github.com/oharar/rbbs) package (@rbbs) and are mentioned in function source code as applicable.
+We especially thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. 
 
 ## References
