### 2019-11-22 dd8edab
Removed tests
