### 2019-11-22 bafd642
Typos and errors
