### 2019-11-22 TrashBirdEcology/bbsAssistant#63 @trashbirdecology
@jsta, in a major revamp of the functionality and the vignette, I have addressed each checked issue you opened above in branch joss-round1. Thanks for the feedback! Will address this in the JOSS review officially when we are ready to resubmit.

The nice-to-have region names has not yet been resolved. Hope to do this soon, though.
