### 2019-11-22 TrashBirdEcology/bbsAssistant#66 @trashbirdecology
@ethanwhite I have simplified subsetting by removing functions eluding to this, and rather supply  examples of species-level (and other taxonomic level) subsetting in the vignette [currently housed in here](https://github.com/TrashBirdEcology/bbsAssistant/blob/master/vignettes/quick-overview.Rmd).

Future development includes more niche species subsetting (e.g., function or argument to remove species with crappy detectability).
