### 2019-11-22 TrashBirdEcology/bbsAssistant#67 @trashbirdecology
Ethan,
I have addressed this through a major functionality revamp, currently in [master branch](https://github.com/TrashBirdEcology/bbsAssistant/tree/master).

The functionality is __much__ more streamlined, and arguably less confusing.

Feather is out. CSV/txt is in.

All functions requiring an internet connection first check to see if the file(s) exist(s) locally, and allow user to specify whether they would like to overwrite with new downloads, or simply import the existing files (using menu). This appears in the "get_" functions.

Thanks again!
