### 2019-11-22 #1768 @TrashBirdEcology
@ethanwhite -- I was trying to work with rdataretriever but couldn't figure it out (zero Py experience here).

To use rdataretriever it seems (please correct me if I am wrong) one must have (a) Python installed and (b) understand how to make R and Python talk. It seems the bbsAssistant package requires only knowledge of R, whereas rdataretriever requires Py & R.

Although I still intend to point out the existence of relevant packages (yours, rbbs, bbsbayes), I am not sure there is too much overlap to go beyond mentioning the package?

Thanks!
