### 2019-11-22 49dfece
Add tests and fix minor bugs and clean up download and import fxn
