### 2019-11-22 eecb535
update travis test skip
