+++ tests/testthat.R
@@ -1,4 +1,7 @@
 library("testthat")
 library("bbsAssistant")
 
-test_check("bbsAssistant")
+# Skip on travis
+if (Sys.getenv("USER") != "travis") {
+ test_check("bbsAssistant")
+}
