### 2019-11-22 879c5f2
Update latter half of paper.md. Add startup message. Fix citations
