+++ paper/paper.bib
@@ -30,8 +30,24 @@
   year={1997}
 }
 
+
+
 @software{rbbs,
   author = {O'Hara, Bob},
   title = {rBBS: R package to import USGS' BBS data},
   url = {https://www.github.com/oharar/rbbs}
 }
+
+@software{bbsBayes,
+  author = {Edwards, Brandon},
+  title = {bbsBayes: An R Package for Hierarchical Bayesian Analysis of North American Breeding Bird Survey Data},
+  url = {https://www.github.com/brandonedwards/bbsbayes}
+}
+
+ @Manual{,
+    title = {rdataretriever: R Interface to the Data Retriever},
+    author = {Daniel McGlinn and Pranita Sharma and David J Harris and Henry Senyondo and Shawn Taylor and Harshit Bansal and Max Pohlman and Ethan White},
+    year = {2019},
+    note = {R package version 2.0.0},
+    url = {https://CRAN.R-project.org/package=rdataretriever},
+  }
