### 2019-11-22 d72a531
minor edits to badges, paper
