+++ paper/paper.bib
@@ -19,7 +19,7 @@
 @article{pardieck2018north,
   title={North American breeding bird survey dataset 1966--2017, version 2017.0},
   author={Pardieck, KL and Ziolkowski Jr, DJ and Lutmerding, M and Hudson, MAR},
-  journal={US Geological Survey, Patuxent Wildlife Research Center, Laurel, Maryland, USA.[online] URL: https://www. pwrc. usgs. gov/BBS/RawData},
+  journal={US Geological Survey, Patuxent Wildlife Research Center, Laurel, Maryland, USA.[online] URL: https://www.pwrc.usgs.gov/BBS/RawData},
   year={2018}
 }
 
@@ -44,7 +44,7 @@
   url = {https://www.github.com/brandonedwards/bbsbayes}
 }
 
- @Manual{,
+ @Manual{rdataretriever,
     title = {rdataretriever: R Interface to the Data Retriever},
     author = {Daniel McGlinn and Pranita Sharma and David J Harris and Henry Senyondo and Shawn Taylor and Harshit Bansal and Max Pohlman and Ethan White},
     year = {2019},
