### 2019-11-22 d31be86
Create CODE_OF_CONDUCT.md
