### 2019-11-22 0f6c4ef
Create CONTRIBUTING.md
