### 2019-11-22 #1768 @TrashBirdEcology
@ethanwhite @jsta  - we have addressed the unchecked boxes in your reviews (in [branch joss-round1](https://github.com/TrashBirdEcology/bbsAssistant/blob/joss-round1/), see below).

We have done some major re-writing of code in /R/, which I hope makes the top-level functioning more intuitive. This is outlined in the [quick overview vignette](https://github.com/TrashBirdEcology/bbsAssistant/blob/joss-round1/vignettes/quick-overview.Rmd).

We have also addressed  the issues you opened in the repository, with the exception of adding mapping features (a feature request by @jsta).

## @ethanwhite
- Community guidelines enhanced by adding a [contributions.md](https://github.com/TrashBirdEcology/bbsAssistant/blob/joss-round1/CONTRIBUTIONS.md) and [codeofconduct.md](https://github.com/TrashBirdEcology/bbsAssistant/blob/joss-round1/CODE_OF_CONDUCT.md).  Will be pushed to master upon resubmission.
-  [Tests](https://github.com/TrashBirdEcology/bbsAssistant/blob/joss-round1/tests/) were updated. Happy to receive suggestions for improving testing.

## @ethanwhite  and @jsta
- A section, State of the Field, and better reference to the relevant packages have been added to the [paper.md](https://github.com/TrashBirdEcology/bbsAssistant/blob/joss-round1/paper/paper.md), and will be regenerated.

We appreciate the time you have already taken to provide critical feedback on this package -- it has been greatly improved.
