+++ paper/paper.bib
@@ -4,7 +4,6 @@
   year={2017}
 }
 
-
 @article{sauer2017first,
   title={The first 50 years of the North American Breeding Bird Survey},
   author={Sauer, John R and Pardieck, Keith L and Ziolkowski Jr, David J and Smith, Adam C and Hudson, Marie-Anne R and Rodriguez, Vicente and Berlanga, Humberto and Niven, Daniel K and Link, William A},
@@ -20,7 +19,7 @@
 @article{pardieck2018north,
   title={North American breeding bird survey dataset 1966--2017, version 2017.0},
   author={Pardieck, KL and Ziolkowski Jr, DJ and Lutmerding, M and Hudson, MAR},
-  journal={US Geological Survey, Patuxent Wildlife Research Center, Laurel, Maryland, USA.[online] URL: https://www. pwrc. usgs. gov/BBS/RawData},
+  journal={US Geological Survey, Patuxent Wildlife Research Center, Laurel, Maryland, USA.[online] URL: https://www.pwrc.usgs.gov/BBS/RawData},
   year={2018}
 }
 
@@ -30,3 +29,25 @@
   journal={Patuxent Wildlife Research Center, Laurel, Maryland},
   year={1997}
 }
+
+
+
+@software{rbbs,
+  author = {O'Hara, Bob},
+  title = {rBBS: R package to import USGS' BBS data},
+  url = {https://www.github.com/oharar/rbbs}
+}
+
+@software{bbsBayes,
+  author = {Edwards, Brandon},
+  title = {bbsBayes: An R Package for Hierarchical Bayesian Analysis of North American Breeding Bird Survey Data},
+  url = {https://www.github.com/brandonedwards/bbsbayes}
+}
+
+ @Manual{rdataretriever,
+    title = {rdataretriever: R Interface to the Data Retriever},
+    author = {Daniel McGlinn and Pranita Sharma and David J Harris and Henry Senyondo and Shawn Taylor and Harshit Bansal and Max Pohlman and Ethan White},
+    year = {2019},
+    note = {R package version 2.0.0},
+    url = {https://CRAN.R-project.org/package=rdataretriever},
+  }
