### 2019-11-22 b5b1188
Delete get_routeInfo.R
