### 2019-11-22 5c6c61b
Update quick-overview.Rmd
