### 2019-11-22 034939d
Delete quick-overview.rmd
