+++ vignettes/quick-overview.rmd
@@ -1,233 +1,0 @@
----
-title: "A quick overview of bbsAssistant functionality"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{vignettes}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
-bibliography: paper/paper.bib
----
-```{r, include = FALSE}
-TRAVIS <- !identical(tolower(Sys.getenv("TRAVIS")), "true")
-knitr::opts_chunk$set(
-  collapse = TRUE,echo=TRUE, warning = FALSE, message=FALSE,
-  comment = "#>",
-  purl = TRAVIS # do not run chunks on TRAVIS
-)
-```
-# Quick Start: Install and Load Development Version and Dependencies
-```{r loadDeps, echo=TRUE, warning=FALSE, message=FALSE}
-devtools::install_github("trashbirdecology/bbsAssistant", force=FALSE)
-library(bbsAssistant)
-```
-
-# Brief Overview of bbsAssistant Functionality 
-This vignette gives a quick overview of the primary functions of the `bbsAssistant` package:
-1. Retrieving, downloading, and manipulating the annually-updated BBS data from the USGS servers via FTP.
-1. Loading and exploring the data stored in the package (updated annually by maintainer or as-needed)
-1. Visualizing BBS data and official BBS model results 
-
-## I: Retrieving, Downloading, and Manipulating route-level BBS data
-This package was creating for expediting the BBS data retrieval process, which currently comprises multiple .zip files retrievable individually via FTP server, and is particularly useful for to those wishing to quickly download and manipulate route-level data from  _multiple regions/states_. 
-
-### Download the BBS data from USGS
-First, please specify a directory within which to store the downloaded BBS data. Here, we specify the directory as `/raw-data/`, and create the directory if it does not already exist. 
-```{r set_dir}
-bbsDir <- here::here("raw-data/")
-dir.create(bbsDir) 
-```
-
-Next, we can use the function, `get_bbsData` to retrieve all or a selection of the data from the USGS webpage using FTP. These compressed (.zip) files are stored to local directory, as specified by argument `data.dir`. Let's download only the data from the U.S. (country code 840) for years 2000-2016:
-
-```{r download_bbs, message=FALSE}
-download_bbs(country.names="USA", data.dir = bbsDir)
-```
-
-### Import the BBS state-level data into R
-Now, let's import the data into R by first decompressing each .zip files of interest from subdirectory `~/raw-data`, and loading each into a list element. Finally, we will condense all into a single data frame. Let's import Florida and Nebraska only:
-```{r import_bbs}
-bbs.df <- import_bbs(data.dir = bbsDir,state.names=c("FlOrIda", "NEBRASKa")) # capitalization irrevlevant.
-# This data frame differs slightly from BBS raw data in that we have appended State Names (`State`) to the df
-bbs.df$State %>% unique()
-```
-
-### Download and route-level information
-Retrieve the route-level information only. This data contains basic information about route location (lat, lon, state, country), activity, BCR location, stratum (BBS), and route name. 
-```{r routes}
-routes <- get_routes(data.dir=bbsDir)
-```
-
-Retrieve the local conditions associated with each route-year iteration. 
-```{r}
-conditions <- get_conditions(data.dir=bbsDir)
-```
-
-### Append taxonomic information and species indices
[…]
