+++ README.Rmd
@@ -26,15 +26,12 @@ library(bbsAssistant)
 ```
 
 ## Function Descriptions and Vignettes
-For function descriptions please build the manual (`devtools::build_manual("bbsAssistant)`) and for an example build the vignette (`usethis::build_vignettes()`; or run `/vignettes/vignettes.Rmd`).
+For function descriptions please build the manual (`devtools::build_manual("bbsAssistant)`) and for an example build the vignette (`devtools::build_vignettes()`; or run `/vignettes/vignettes.Rmd`).
 
 ## Contributions
 To make a contribution visit the [contributions instructions](https://trashbirdecology.github.io/bbsAssistant/CONTRIBUTING.html). Contributors must adhere to the [Code of Conduct](https://trashbirdecology.github.io/bbsAssistant/CODE_OF_CONDUCT.html). 
 
 ## Acknowledgments
-
-We especially thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. We also thank [Ethan White; \@ethanwhite](https://github.com/ethanwhite) and [Joseph Stachelek; \@jsta](https://github.com/ethanwhite) for their feedback on a previous iteartion of this software which greatly improved usability, functionality, and efficiency. We also thank past and future contributors to this package. 
-
-We thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. Some functions in this package were adapted from the [rBBS](https://github.com/oharar/rbbs) package and are mentioned in function source code as applicable.
+We especially thank the participatory scientists who collect data annually for the North American Breeding Bird Survey, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible. 
 
 ## References
