### 2019-11-24 5e357cf
Delete LICENSE.md
Public domain.
