### 2019-11-24 7a72eaa
Add CCO license
