### 2019-11-24 570efc5
Add CCO license file
