+++ LICENSE.md
@@ -1,0 +1,8 @@
+R Package `bbsAssistant`
+
+To the extent possible under law, the person who associated CC0 with
+R Package `bbsAssistant` has waived all copyright and related or neighboring rights
+to R Package `bbsAssistant`.
+
+Please see <http://creativecommons.org/publicdomain/zero/1.0/> for a copy of the CC0 legalcode along with this
+work.  
