### 2019-11-25 ff750f9
Add a wrapper function for downloading and importing.
