+++ NAMESPACE
@@ -8,5 +8,6 @@ export(get_credibility_trends)
 export(get_routes)
 export(get_species)
 export(import_bbs)
+export(quick_get_bbs)
 importFrom(magrittr,"%>%")
 importFrom(utils,download.file)
