### 2019-11-25 13034fb
add wrapper fxn for downloading, importing, and merging df with weather conditions for quick use
