+++ R/get_routes.R
@@ -5,8 +5,9 @@
 #' @param route.type Default = 1 (roadside surveys only). One or more numbers indicating route substrate (1=roadside;2=water;3=off-road; Default = 1, roadside only).
 #' @param bbs.stratum Vector of BBS physiographic stratum codes by which to filter the routes.
 #' @param bcr.num A vector of Bird Conservation Region codes where by which to filter the routes.
-#' #' @return A dataframe containing route-level information. 
+#' @return A dataframe containing route-level information. 
 #' @param active.only Logical. Default=TRUE. Keep only active routes (ie.. Active==1). 
+#' @param overwrite Logical. Default NULL. Will force overwrite existing file named "routes.csv" if it exists in data.dir. If NULL will prompt user to specify overwrite. 
 #' @importFrom magrittr %>%
 #' @export get_routes
 #' @examples
@@ -22,17 +23,22 @@ get_routes <- function(
     active.only = TRUE,
     route.type = 1,
     bbs.stratum = NULL,
-    bcr.num = NULL
+    bcr.num = NULL, 
+    overwrite=NULL
     ) {
     
     suppressWarnings(dir.create(data.dir))
 
     # Prompt to overwrite the routes data
-    if("routes.csv" %in% list.files(data.dir)){
+    if(is.null(overwrite) & "routes.csv" %in% list.files(data.dir)){
        choice <- menu(c("Yes, download most recent version of routes.csv", "No, import existing routes.csv from local directory "))
-    }else(choice=1)
+       if(choice==1) overwrite <- TRUE
+       if(choice==2) overwrite <- FALSE
+       rm(choice)
+    }
+
         
-     if(choice==1){ # if user wants to overwrite OR data DNE then download and unzip to file in data.dir
+    if(overwrite | !("routes.csv" %in% list.files(data.dir))){ # if user wants to overwrite OR data DNE then download and unzip to file in data.dir
     # specify where routes.zip will be saved
     routes.local.path = paste0(data.dir, "/routes.zip")
     
@@ -45,12 +51,13 @@ get_routes <- function(
    
 
     # read in the unzipped file
-    routes.df <- read.csv(list.files(data.dir, pattern="routes.csv", full.names = TRUE))
+    routes <- read.csv(list.files(data.dir, pattern="routes.csv", full.names = TRUE))
     
     
     # Join region codes to the route information
     if(!exists("region_codes")) data(region_codes) # laod region_codes
-    routes <- merge(routes.df, region_codes ) %>% dplyr::select(-zip_states)
+    
+    routes <- merge(routes, region_codes ) %>% dplyr::select(-zip_states)
     
     ##### FILTER THE ROUTES #######
     # Keep only active routes
