### 2019-11-25 93d78ca
Update license and minor readme and description edits
