### 2019-11-25 0e47193
Update README and minor quick_get_bbs edit
