+++ README.Rmd
@@ -10,12 +10,10 @@ output: github_document
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis build status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>[![License: CC0-1.0](https://img.shields.io/badge/License-CC0%201.0-lightgrey.svg)](http://creativecommons.org/publicdomain/zero/1.0/)
 
 ## About
+This R package contains functions for downloading and munging data from the [North American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/) (BBS) [via FTP](https://www.pwrc.usgs.gov/BBS/RawData/) [@pardieck2018north; @sauer2017first]. This package was created to allow the user to bulk-download the BBS point count and related (e.g., route-level conditions) via FTP, and to quickly subset the data by taxonomic classifications and/or geographical locations. This package also maintains data containing the trend and annual indices from the most recent (1996-2017) [hierarhchical popultion trend analyses](https://www.mbr-pwrc.usgs.gov/bbs/) [@sauer2017north]. 
 
-This package contains functions for downloading and munging data from the North American Breeding Bird Survey (BBS) FTP server [@pardieck2018north; @sauer2017first]. Although the FTP server provides a public interface for retrieving data and analysis results, this package consolidates the efforts of the data user by automating downloading and decompression of .zip data files, downloading route-level information, and saving them as .feather files for speedy import from disk. The data subsetting features of this package also allow the user to readily import and save to file only the data necessary for her purposes. Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
-
-The BBS team uses hierarchical modelling techniques to generate population trend estimates [@sauer2017north] at various spatial scales [see the BBS results webpage](https://www.mbr-pwrc.usgs.gov/). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package contains two functions for retrieving the population trend estimates produced by @sauer2017north and the associated data credibility scores: a web-scraping function for obtaining current region and/or species-specific population trend estimates and data credibility scores via a supplied url, [`get_credibility_trends()`](https://github.com/TrashBirdEcology/bbsAssistant/blob/master/R/get_credibility_trends.R); and a function for the current and archived population trends estimates for *all* species and regions, [`get_analysis_results()`](https://github.com/TrashBirdEcology/bbsAssistant/blob/master/R/get_analysis_results.R). Further, the package contains data objects of these analysis results, and can be retrieved using the function utils::data(). Call `data(package="bbsAssistant")` for data objects and descriptions.
-
-## Installing `bbsAssistant`
+## Quick Start
+### Installing __bbsAssistant__
 Install the development version of this package using devtools and load the package and dependencies:
 ```{r install, echo=TRUE, warning=FALSE, message=FALSE, eval=FALSE}
 devtools::install_github("trashbirdecology/bbsAssistant", 
@@ -23,24 +21,23 @@ devtools::install_github("trashbirdecology/bbsAssistant",
 library(bbsAssistant)
 ```
 
-## Quick Start
-Quickly retrieve all or a subset (of states/regions) of the BBS data using the wrapper function, ``bbsAssistant::quick_get_bbs()`:
-```{r quick-get-bbs}
-bbs <- quick_get_bbs(state.names = c("Florida", "Nebraska"),  # get only two states for convenience. Leave blank to retrieve all states/regions.
+## Quick-download and import
[…]
