+++ README.md
@@ -1,51 +1,33 @@
 **bbsAssistant**: An R package for downloading and handling data and
 information from the North American Breeding Bird Survey.
 ================
-Last updated: 2019-11-22
+Last updated: 2019-11-25
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
 [![status](https://joss.theoj.org/papers/4b445373a7a7806c92e17bdd194a8e69/status.svg)](https://joss.theoj.org/papers/4b445373a7a7806c92e17bdd194a8e69)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis
 build
-status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>
-<!-- [![Coverage status](https://codecov.io/gh/trashbirdecology/bbsAssistant/master/graph/badge.svg)](https://codecov.io/github/trashbirdecology/bbsAssistant?branch=master) -->
-[![License: CC0-1.0](https://img.shields.io/badge/License-CC0%201.0-lightgrey.svg)](http://creativecommons.org/publicdomain/zero/1.0/)
+status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>[![License:
+CC0-1.0](https://img.shields.io/badge/License-CC0%201.0-lightgrey.svg)](http://creativecommons.org/publicdomain/zero/1.0/)
 
 ## About
 
-This package contains functions for downloading and munging data from
-the North American Breeding Bird Survey (BBS) FTP server (Pardieck et
-al. 2018; J. R. Sauer et al. 2017). Although the FTP server provides a
-public interface for retrieving data and analysis results, this package
-consolidates the efforts of the data user by automating downloading and
-decompression of .zip data files, downloading route-level information,
-and saving them as .feather files for speedy import from disk. The data
-subsetting features of this package also allow the user to readily
-import and save to file only the data necessary for her purposes.
-Although the primary audience is for those wishing to use BBS data in
-Program R for visualization or analysis, this package can be used to
-quickly download the BBS data to file for use elsewhere.
+This R package contains functions for downloading and munging data from
+the [North American Breeding Bird
+Survey](https://www.pwrc.usgs.gov/bbs/) (BBS) [via
+FTP](https://www.pwrc.usgs.gov/BBS/RawData/) (Pardieck et al. 2018; J.
+R. Sauer et al. 2017). This package was created to allow the user to
+bulk-download the BBS point count and related (e.g., route-level
+conditions) via FTP, and to quickly subset the data by taxonomic
+classifications and/or geographical locations. This package also
+maintains data containing the trend and annual indices from the most
+recent (1996-2017) [hierarhchical popultion trend
+analyses](https://www.mbr-pwrc.usgs.gov/bbs/) (J. Sauer et al. 2017).
 
-The BBS team uses hierarchical modelling techniques to generate
-population trend estimates (J. Sauer et al. 2017) at various spatial
-scales [see the BBS results webpage](https://www.mbr-pwrc.usgs.gov/).
-Given the variability in data availability, the BBS team also provides
-data credibility scores for species-regions combinations. This package
-contains two functions for retrieving the population trend estimates
-produced by J. Sauer et al. (2017) and the associated data credibility
-scores: a web-scraping function for obtaining current region and/or
-species-specific population trend estimates and data credibility scores
-via a supplied url,
[…]
