### 2019-11-25 f1b3156
Update Vignette to reflefct functional changes.
