### 2019-11-25 f27525c
Update and simplify the paper.md
