### 2019-11-26 #1768 @TrashBirdEcology
Thanks, @ethanwhite!

Jessica L. Burnett

GitHub <http://github.com/trashbirdecology>
CV/Resume
<https://github.com/TrashBirdEcology/cv/blob/master/burnett_extendedCV.pdf>
ORC ID: 0000-0002-0896-5099 <https://orcid.org/0000-0002-0896-5099>

On Tue, Nov 26, 2019 at 7:06 AM Ethan White <notifications@github.com>
wrote:

> @TrashBirdEcology <https://github.com/TrashBirdEcology> - apologies for
> the difficulties with rdataretriever (don't see them in this issue any
> more, but got a email notification about them). The retriculate package
> that we rely on for talking to Python has introduced a number of bugs and
> we're in the midst of trying to mitigate these issues through improved
> installation instructions until they get them fixed. You current mention of
> this and other associated packages is all that's needed here.
>
> @TrashBirdEcology <https://github.com/TrashBirdEcology> @kthyng
> <https://github.com/kthyng> - these changes all look great and I very
> much appreciate the productive engagement both on these issues and
> suggestions related to design! It's been a pleasure working with you both.
> I've completed the rest of the check boxes and consider this software paper
> ready for publication.
>
