### 2019-11-26 2fcc08b
Add force=TRUE  to install
