### 2019-11-26 f11d890
Add force=TRUE to install
