### 2019-11-26 4b0a76f
Update badges; add contributors section
