+++ paper/paper.md
@@ -39,6 +39,6 @@ The Patuxent Wildlife Research Center uses hierarchical modelling of the BBS dat
 We are aware of three R packages which retrieve and/or munge the NABBS data: __rdataretriever__ [@rdataretriever], __rBBS__ [@rBBS], and __bbsBayes__ [@bbsBayes], each of which provides various pathways for importing NABBS data into the local environment. rdataretriever provides a dataset which integrates components of the BBS data, however, requires the use of Python [@python] in conjunction with R. The __rBBS__ package is perhaps most aligned with the __bbsAssistant__ package, in that it also provides functions for downloading NABBS data, however, the existing repository is apparently stale. Furhter, bbsAssistant provides streamlined functionality for retrieving location and species-specific data. Finally, the bbsBayes package, was creating primarily to run hierarhical models in a Bayesian framework within R. __bbsBayes__ provides a function (`bbsBayes::fetch_bbs_data()`) for retrieving and importing all BBS data, yet does not currently allow for custom download and importation. 
  
 # Acknowledgements
-We especially thank the thousands of participatory scientists who have and continuie to collect data for the North American Breeding Bird Survey program, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible.  We also thank [Ethan White; \@ethanwhite](https://github.com/ethanwhite) and [Joseph Stachelek; \@jsta](https://github.com/ethanwhite) for their feedback on a previous iteartion of this software which greatly improved usability, functionality, and efficiency. This draft manuscript is distributed solely for purposes of scientific peer review. Its content is deliberative and predecisional, so it must not be disclosed or released by reviewers. Because the manuscript has not yet been approved for publication by the U.S. Geological Survey (USGS), it does not represent any official USGS finding or policy. This software is preliminary or provisional and is subject to revision. It is being provided to meet the need for timely best science. The software has not received final approval by the U.S. Geological Survey (USGS). No warranty, expressed or implied, is made by the USGS or the U.S. Government as to the functionality of the software and related material nor shall the fact of release constitute any such warranty. The software is provided on the condition that neither the USGS nor the U.S. Government shall be held liable for any damages resulting from the authorized or unauthorized use of the software.
+We especially thank the thousands of participatory scientists who have and continuie to collect data for the North American Breeding Bird Survey program, and the Patuxent Wildlife Research Center for making these data publicly and easily accessible.  We also thank [Ethan White; \@ethanwhite](https://github.com/ethanwhite) and [Joseph Stachelek; \@jsta](https://github.com/ethanwhite) for their feedback on a previous iteration of this software which greatly improved usability, functionality, and efficiency. This draft manuscript is distributed solely for purposes of scientific peer review. Its content is deliberative and predecisional, so it must not be disclosed or released by reviewers. Because the manuscript has not yet been approved for publication by the U.S. Geological Survey (USGS), it does not represent any official USGS finding or policy. This software is preliminary or provisional and is subject to revision. It is being provided to meet the need for timely best science. The software has not received final approval by the U.S. Geological Survey (USGS). No warranty, expressed or implied, is made by the USGS or the U.S. Government as to the functionality of the software and related material nor shall the fact of release constitute any such warranty. The software is provided on the condition that neither the USGS nor the U.S. Government shall be held liable for any damages resulting from the authorized or unauthorized use of the software.
 
 # References
