### 2019-11-26 720952b
Update contributors, team sections, add contr. # badge.
