+++ README.md
@@ -6,7 +6,9 @@ Last updated: 2019-11-26
 <!-- README.md is generated from README.Rmd. Please edit that file and render to push updates.-->
 
 [![status](https://joss.theoj.org/papers/4b445373a7a7806c92e17bdd194a8e69/status.svg)](https://joss.theoj.org/papers/4b445373a7a7806c92e17bdd194a8e69)[![License:
-CC0-1.0](https://img.shields.io/badge/License-CC0%201.0-lightgrey.svg)](http://creativecommons.org/publicdomain/zero/1.0/)[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-lightgrey.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis
+CC0-1.0](https://img.shields.io/badge/License-CC0%201.0-lightgrey.svg)](http://creativecommons.org/publicdomain/zero/1.0/)[![All
+Contributors](https://img.shields.io/badge/all_contributors-4-lightgrey.svg?style=flat-square)](#contributors)
+[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-lightgrey.svg)](https://www.tidyverse.org/lifecycle/#maturing)[![Travis
 build
 status](https://travis-ci.org/trashbirdecology/bbsAssistant.svg?branch=master)](https://travis-ci.org/trashbirdecology/bbsAssistant)<img src="man/figures/logo.png" align="right" height=140/>
 
@@ -70,14 +72,42 @@ For questions, comments, or issues, feel free to email the maintainer
 [Issue](https://github.com/TrashBirdEcology/bbsAssistant/issues)
 (preferred).
 
-## Contributors ✨
+## Project Team
 
-Thanks goes to these wonderful people ([emoji
+<table>
+
+<tr>
+
+<td align="center">
+
+<a href="http://trashbirdecology.github.io/"><img src="https://avatars2.githubusercontent.com/u/9939381?s=460&v=4" width="100px;" alt="Jessica Burnett"/><br /><sub><b>Jessica
+Burnett <br>Team Lead &
+Maintainer</b></sub></a><br />
+
+<td align="center">
+
+<a href="https://github.com/GabsPalomo"><img src="https://avatars1.githubusercontent.com/u/28967490?s=460&v=4" width="100px;" alt="Gabby Palomo-Muñoz"/><br /><sub><b>Gabby
+Palomo-Muñoz <br>Team
+Member</b></sub></a><br />
+
+</td>
+
+<td align="center">
+
+<a href="https://github.com/lsw5077"><img src="https://avatars0.githubusercontent.com/u/22730128?s=460&v=4" width="100px;" alt="Lyndsie Wszola"/><br /><sub><b>Lyndsie
+Wszola <br>Team Member</b></sub></a><br />
+
+</td>
+
+</tr>
+
+</table>
+
+## Contributors
+
+Thanks to all contributors ([emoji
 key](https://allcontributors.org/docs/en/emoji-key)):
-
-<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
-
-<!-- prettier-ignore -->
+<!-- ALL-CONTRIBUTORS-LIST:START-->
 
 <table>
 
@@ -86,7 +116,35 @@ key](https://allcontributors.org/docs/en/emoji-key)):
 <td align="center">
 
 <a href="http://ethanwhite.org"><img src="https://avatars0.githubusercontent.com/u/744427?v=4" width="100px;" alt="Ethan White"/><br /><sub><b>Ethan
-White</b></sub></a><br /><a href="#review-ethanwhite" title="Reviewed Pull Requests">👀</a>
+White</b></sub></a><br />
[…]
