### 2019-11-26 #1768 @jsta
Looks good to me. The issues I raised with the initial review have all been addressed. Only thing I see remaining is that the paper still mentions the `feather` file format even though this functionality has been removed?
