### 2019-11-26 2e94a5c
Remove contribution emojis
Just list names, avatars, websites.
