### 2019-11-26 f293e84
remove all-congtributors specification
