+++ README.md
@@ -154,10 +154,6 @@ Gruson</b></sub></a>
 
 <!-- ALL-CONTRIBUTORS-LIST:END -->
 
-This project follows the
-[all-contributors](https://github.com/all-contributors/all-contributors)
-specification. Contributions of any kind welcome\!
-
 ## Acknowledgments
 
 We especially thank the participatory scientists who collect data
