### 2019-11-26 #1768 @TrashBirdEcology
Yikes -- I guess I didn't push that edit.

This has been resolved @jsta.

Jessica L. Burnett

GitHub <http://github.com/trashbirdecology>
CV/Resume
<https://github.com/TrashBirdEcology/cv/blob/master/burnett_extendedCV.pdf>
ORC ID: 0000-0002-0896-5099 <https://orcid.org/0000-0002-0896-5099>

On Tue, Nov 26, 2019 at 10:04 AM Joseph Stachelek <notifications@github.com>
wrote:

> Looks good to me. The issues I raised with the initial review have all
> been addressed. Only thing I see remaining is that the paper still mentions
> the feather file format even though this functionality has been removed?
>
