### 2019-11-26 7c78119
typo
Fix typos per @chrischizinski
