### 2019-11-26 84d9451
Fix bug - get_species
Fix bug in issue [83](https://github.com/TrashBirdEcology/bbsAssistant/issues/83)--get_species.R

Also ignore files in /raw-data
