### 2019-11-26 a5d391f
Fix namespace to include export fxns
fix: run load_all(), then build, then document().
