+++ NAMESPACE
@@ -1,4 +1,13 @@
 # Generated by roxygen2: do not edit by hand
 
+export("%>%")
+export(download_bbs)
+export(get_analysis_results)
+export(get_conditions)
+export(get_credibility_trends)
+export(get_routes)
+export(get_species)
+export(import_bbs)
+export(quick_get_bbs)
 importFrom(magrittr,"%>%")
 importFrom(utils,download.file)
