### 2019-11-26 ef12c49
Minor edits
