### 2019-11-26 f7041c6
Reduce downlaod filesize in vignette example of dwonload_bbs
