### 2019-11-26 cd98f35
Minor writing updates + fix .bib bug in vignette
Also fixed text in section ### Download the BBS data from USGS

bug per chizinski -- thanks!
