### 2019-11-26 ea286d3
Do not eval get_analysis_results chunk
Having issues rendering the vignette when specifying a .bib file. Still working that one out
