### 2019-11-26 8584cdf
Add bib to proj dir
