+++ bib.bib
@@ -1,0 +1,53 @@
+@article{sauer2017north,
+  title={The North American breeding bird survey, results and analysis 1966-2015},
+  author={Sauer, JR and Niven, D and Hines, J and Ziolkowski Jr, David and Pardieck, KL and Fallon, JE and Link, William},
+  year={2017}
+}
+
+@article{sauer2017first,
+  title={The first 50 years of the North American Breeding Bird Survey},
+  author={Sauer, John R and Pardieck, Keith L and Ziolkowski Jr, David J and Smith, Adam C and Hudson, Marie-Anne R and Rodriguez, Vicente and Berlanga, Humberto and Niven, Daniel K and Link, William A},
+  journal={The Condor: Ornithological Applications},
+  volume={119},
+  number={3},
+  pages={576--593},
+  year={2017},
+  publisher={Oxford University Press},
+  doi={10.1650/CONDOR-17-83.1}
+}
+
+@article{pardieck2018north,
+  title={North American breeding bird survey dataset 1966--2017, version 2017.0},
+  author={Pardieck, KL and Ziolkowski Jr, DJ and Lutmerding, M and Hudson, MAR},
+  journal={US Geological Survey, Patuxent Wildlife Research Center, Laurel, Maryland, USA.[online] URL: https://www.pwrc.usgs.gov/BBS/RawData},
+  year={2018}
+}
+
+@article{sauer1997north,
+  title={The North American breeding bird survey results and analysis, version 96.4},
+  author={Sauer, JR and Hines, JE and Gough, G and Thomas, I and Peterjohn, BG},
+  journal={Patuxent Wildlife Research Center, Laurel, Maryland},
+  year={1997}
+}
+
+
+
+@software{rbbs,
+  author = {O'Hara, Bob},
+  title = {rBBS: R package to import USGS' BBS data},
+  url = {https://www.github.com/oharar/rbbs}
+}
+
+@software{bbsBayes,
+  author = {Edwards, Brandon},
+  title = {bbsBayes: An R Package for Hierarchical Bayesian Analysis of North American Breeding Bird Survey Data},
+  url = {https://www.github.com/brandonedwards/bbsbayes}
+}
+
+ @Manual{rdataretriever,
+    title = {rdataretriever: R Interface to the Data Retriever},
+    author = {Daniel McGlinn and Pranita Sharma and David J Harris and Henry Senyondo and Shawn Taylor and Harshit Bansal and Max Pohlman and Ethan White},
+    year = {2019},
+    note = {R package version 2.0.0},
+    url = {https://CRAN.R-project.org/package=rdataretriever},
+  }
