+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'bbsAsssistant: A package for downloading and munging data and information from the North American Breeding Bird Survey'
+title: 'bbsAssistant: A package for downloading and munging data and information from the North American Breeding Bird Survey'
 authors:
 - affiliation: '1'
   name: Jessica L Burnett
@@ -27,16 +27,16 @@ affiliations:
 ---
 
 # Summary
-This R package contains functions for downloading and munging data from the U.S. Geological Survey's [North American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/) (BBS) [via file  transfer protocol (FTP)](https://www.pwrc.usgs.gov/BBS/RawData/) [@pardieck2018north; @sauer2017first]. This package was created to allow the user to bulk-download the BBS point count and related (e.g., route-level conditions) via FTP, and to quickly subset the data by taxonomic classifications and/or geographical locations. This package also maintains data containing the trend and annual indices from the most recent (1996-2017) [hierarhchical popultion trend analyses](https://www.mbr-pwrc.usgs.gov/bbs/) [@sauer2017north]. 
+This R package contains functions for downloading and munging data from the U.S. Geological Survey's [North American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/) (BBS) [via file  transfer protocol (FTP)](https://www.pwrc.usgs.gov/BBS/RawData/) [@pardieck2018north; @sauer2017first]. This package was created to allow the user to bulk-download the BBS point count and related (e.g., route-level conditions) via FTP, and to quickly subset the data by taxonomic classifications and/or geographical locations. This package also maintains data containing the trend and annual indices from the most recent (1996-2017) [hierarchical population analyses](https://www.mbr-pwrc.usgs.gov/bbs/) [@sauer2017north]. 
 
 ## Retrieving and Munging Point-count and Related Data
-Although the BBS provides a public interface for retrieving data and analysis results via FTP, __bbsAsssitant__ expedites the efforts of downloading, decompressing, importing, and subsetting the state/region-level point-count files, and other associated files (e.g., taxonomic information, geographic information, route-level conditions, geographical information). The data subsetting features allow the user to download and import only the data necessary for their purposes, rather than the entire data collection. Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
+Although the BBS provides a public interface for retrieving data and analysis results via FTP, __bbsAssistant__ expedites the efforts of downloading, decompressing, importing, and subsetting the state/region-level point-count files, and other associated files (e.g., taxonomic information, geographic information, route-level conditions, geographical information). The data subsetting features allow the user to download and import only the data necessary for their purposes, rather than the entire data collection. Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
 
 ## Retrieving Population Trend Model Results from the BBS website
 The Patuxent Wildlife Research Center uses hierarchical modelling of the BBS data to generate population trend estimates and annual indices at various spatial scales [@sauer2017north; see also the BBS results webpage](https://www.mbr-pwrc.usgs.gov/). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package contains the most recent results associated with these analyses as data objects, but also provides a function (`get_credibility_trends()`) for retrieving all analysis results which are public-facing. 
 
 # State of the Field
[…]
