+++ paper/paper.md
@@ -9,6 +9,7 @@ authors:
   orcid: 0000-0002-2660-2048
 - affiliation: 2
   name: Gabriela Palomo-Muñoz
+  orcid: 0000-0002-0661-413X
 date: "26 November 2019"
 output: pdf_document
 bibliography: paper.bib
