### 2019-11-26 5fd3bba
Update function-overview.Rmd
