+++ vignettes/function-overview.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "A quick overview of bbsAssistant functionality"
 output: rmarkdown::html_vignette
-# bibliography: "bib.bib"
+bibliography: "bib.bib"
 vignette: >
   %\VignetteIndexEntry{vignettes}
   %\VignetteEngine{knitr::rmarkdown}
