### 2019-11-26 9a01509
Remove emoji key
