### 2019-11-26 #1768 @kthyng
Great! I see the go-ahead from @ethanwhite.

@jsta can you confirm that the recent change has resolved your last concern?
