### 2019-12-02 #1768 @kthyng
Great! We can begin the final acceptance process then.
