### 2019-12-02 944369a
Fixes to a few link/reference combos
Also my name :)
