+++ paper/paper.md
@@ -34,12 +34,12 @@ This R package contains functions for downloading and munging data from the U.S.
 Although the BBS provides a public interface for retrieving data and analysis results via FTP, __bbsAssistant__ expedites the efforts of downloading, decompressing, importing, and subsetting the state/region-level point-count files, and other associated files (e.g., taxonomic information, geographic information, route-level conditions, geographical information). The data subsetting features allow the user to download and import only the data necessary for their purposes, rather than the entire data collection. Although the primary audience is for those wishing to use BBS data in Program R for visualization or analysis, this package can be used to quickly download the BBS data to file for use elsewhere. 
 
 ## Retrieving Population Trend Model Results from the BBS website
-The Patuxent Wildlife Research Center uses hierarchical modelling of the BBS data to generate population trend estimates and annual indices at various spatial scales [@sauer2017north; see also the BBS results webpage](https://www.mbr-pwrc.usgs.gov/). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package contains the most recent results associated with these analyses as data objects, but also provides a function (`get_credibility_trends()`) for retrieving all analysis results which are public-facing. 
+The Patuxent Wildlife Research Center uses hierarchical modelling of the BBS data to generate population trend estimates and annual indices at various spatial scales (@sauer2017north; see also the [BBS results webpage](https://www.mbr-pwrc.usgs.gov/)). Given the variability in data availability, the BBS team also provides data credibility scores for species-regions combinations. This package contains the most recent results associated with these analyses as data objects, but also provides a function (`get_credibility_trends()`) for retrieving all analysis results which are public-facing. 
 
 # State of the Field
 We are aware of three R packages which retrieve and/or munge the BBS data: __rdataretriever__ [@rdataretriever], __rBBS__ [@rbbs], and __bbsBayes__ [@bbsBayes], each of which provides various pathways for importing BBS data into the local environment. rdataretriever provides a data set which integrates components of the BBS data, however, requires the use of Python in conjunction with R. The __rBBS__ package is perhaps most aligned with the __bbsAssistant__ package, in that it also provides functions for downloading BBS data, however, the existing repository is apparently stale. Further, bbsAssistant provides streamlined functionality for retrieving location and species-specific data. Finally, the bbsBayes package, was creating primarily to run hierarchical models in a Bayesian framework within R. __bbsBayes__ provides a function (`bbsBayes::fetch_bbs_data()`) for retrieving and importing all BBS data, yet does not currently allow for custom download and importation. 
  
 # Acknowledgements
[…]
