### 2019-12-02 #1768 @kthyng
I read through the paper and made a few small fixed in [this PR](https://github.com/TrashBirdEcology/bbsAssistant/pull/85) if you want to merge.
