### 2019-12-02 #1768 @kthyng
Forgot to look through your references before I made that PR so here is another: https://github.com/TrashBirdEcology/bbsAssistant/pull/86
