### 2019-12-02 #1768 @kthyng
Sorry I should have been more specific, by "archive" I mean to submit your JOSS submission to Zenodo or similar for a DOI.
