### 2019-12-02 #1768 @TrashBirdEcology
@kthyng, yes, I will archive on USGS, but that might take a few days. I will ping you as soon as this is complete.
