### 2019-12-02 #1768 @kthyng
When you archive with USGS will it be the code and paper like you would with Zenodo? And you get a DOI?

Also in case it was lost in the mix, please do make a tagged release and we'll set that as the new version. Your original version number is ok but it should match a tagged release in github.
