### 2019-12-02 #1768 @TrashBirdEcology
Yes, similar to Zenodo, including a DOI.

Not sure what is meant by 'tagged' release, however?

Jessica L. Burnett

GitHub <http://github.com/trashbirdecology>
CV/Resume
<https://github.com/TrashBirdEcology/cv/blob/master/burnett_extendedCV.pdf>
ORC ID: 0000-0002-0896-5099 <https://orcid.org/0000-0002-0896-5099>

On Mon, Dec 2, 2019 at 8:15 AM Kristen Thyng <notifications@github.com>
wrote:

> When you archive with USGS will it be the code and paper like you would
> with Zenodo? And you get a DOI?
>
> Also in case it was lost in the mix, please do make a tagged release and
> we'll set that as the new version. Your original version number is ok but
> it should match a tagged release in github.
>
