### 2019-12-02 #1768 @kthyng
@TrashBirdEcology I'm following editor guidelines here: https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance

And here is some information about making releases in github: https://help.github.com/en/github/administering-a-repository/creating-releases
