### 2019-12-02 #1768 @kthyng
@TrashBirdEcology Looks like the changes to the references haven't made it through?
