### 2019-12-02 #1768 @TrashBirdEcology
@kthyng  the
__DOI__:  10.5066/P93W0EAW (will take a couple of hours to be public).
__Version release__: https://github.com/TrashBirdEcology/bbsAssistant/releases/tag/v1.0

DOI: 10.5066/P93W0EAW
