+++ README.md
@@ -166,7 +166,7 @@ of all sizes to this project. Finally, we thank two peer reviewers,
 [Ethan White](www.github.com/ethanwhite) and [Josepha
 Stachelek](www.github.com/jsta) whose feedback greatly improved the
 quality of this software and the [associated software
-paper](www.github.com/trashbirdecology/bbsassistant/paper/paper.md).
+paper](www.github.com/trashbirdecology/bbsassistant/paper/paper.md). [Logo](https://github.com/TrashBirdEcology/bbsAssistant/blob/master/man/figures/logo.png) by Gabby Palomo-Munoz.
 This software is preliminary or provisional and is subject to revision.
 It is being provided to meet the need for timely best science. The
 software has not received final approval by the U.S. Geological Survey
@@ -175,7 +175,7 @@ U.S. Government as to the functionality of the software and related
 material nor shall the fact of release constitute any such warranty. The
 software is provided on the condition that neither the USGS nor the U.S.
 Government shall be held liable for any damages resulting from the
-authorized or unauthorized use of the software.
+authorized or unauthorized use of the software. 
 
 ## References
 
