### 2019-12-02 c969b82
Add USGS biolab link + logo design credit
