### 2019-12-02 91ba189
Add USGS software release citation
