### 2019-12-02 c781650
Add USGs software release citation
