### 2019-12-03 #1768 @kthyng
@TrashBirdEcology I'm not familiar with this archive setup. Normally with Zenodo I would go to the website of your submission there and make sure the title and author list match your paper. Can you do that for your archive?
