### 2019-12-03 #1768 @TrashBirdEcology
@kthyng I apologies for being a difficult customer here, but I am not sure what you mean by this?  The DOI archive (https://doi.org/10.5066/P93W0EAW) contains a .zip file of the V1.0 release instance of the package.
