### 2019-12-04 #1768 @kthyng
@TrashBirdEcology You're not being difficult at all! I see at your archive link now that your title and author list is consistent with your JOSS submission. I wanted to check that since it isn't always true.
