### 2019-12-04 #1768 @TrashBirdEcology
Ahh yes I see you referred to the title and authorship discrepancies. Apologies!!

Sent from my iPhone

> On Dec 4, 2019, at 7:50 AM, Kristen Thyng <notifications@github.com> wrote:
>
> ﻿
> @TrashBirdEcology You're not being difficult at all! I see at your archive link now that your title and author list is consistent with your JOSS submission. I wanted to check that since it isn't always true.
>
