### 2019-12-04 #1768 @kthyng
Congratulations on your new paper @TrashBirdEcology!!! Many thanks to reviewers @ethanwhite and @jsta for your time and expertise.
