### 2019-12-04 #1768 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
